tests: Only use new e2e, and remove old e2e tests
