+++ test/e2e/aws_test.go
@@ -1,7 +1,7 @@
 // +build e2e
 
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2020 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,65 +16,40 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_test
+package e2e
 
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
-	"os/exec"
-	"path"
+	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
 
-	"os" // used in watchEvents
-
-	apimachinerytypes "k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/pointer"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/aws/client"
+	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/elb"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"github.com/gophercloud/gophercloud"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	apirand "k8s.io/apimachinery/pkg/util/rand"
-	"k8s.io/client-go/informers"   // used in watchEvents
-	"k8s.io/client-go/tools/cache" // used in watchEvents
-	"k8s.io/client-go/tools/clientcmd"
+	apimachinerytypes "k8s.io/apimachinery/pkg/types"
+	"k8s.io/utils/pointer"
+	infrav1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha3"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1alpha3"
 	kubeadmv1beta1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/v1beta1"
-	capiFlag "sigs.k8s.io/cluster-api/test/helpers/flag"
+	"sigs.k8s.io/cluster-api/test/framework"
+	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 	"sigs.k8s.io/cluster-api/util"
-	"sigs.k8s.io/cluster-api/util/kubeconfig"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
-
-	infrav1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha3"
-)
-
-var (
-	cniManifests      = capiFlag.DefineOrLookupStringFlag("cniManifests", "https://docs.projectcalico.org/manifests/calico.yaml", "URL to CNI manifests to load")
-	kubectlBinary     = capiFlag.DefineOrLookupStringFlag("kubectlBinary", "kubectl", "path to the kubectl binary")
-	availabilityZones []*string
-	privateCIDRs      = [3]string{"10.0.0.0/24", "10.0.2.0/24", "10.0.4.0/24"}
-	publicCIDRs       = [3]string{"10.0.1.0/24", "10.0.3.0/24", "10.0.5.0/24"}
-	tries             = 0
-)
-
-const (
-	scaleUpReplicas   = 5
-	scaleDownReplicas = 0
 )
 
 type statefulSetInfo struct {
@@ -94,127 +69,20 @@ type statefulSetInfo struct {
 	volMountPath              string
 }
 
-type testSetup struct {
-	namespace               string
-	clusterName             string
-	awsClusterName          string
-	cpMachinePrefix         string
-	cpAWSMachinePrefix      string
-	cpBootstrapConfigPrefix string
-	mdBootstrapConfig       string
-	machineDeploymentName   string
-	awsMachineTemplateName  string
-	testTmpDir              string
-	instanceType            string
-	initialReplicas         int32
-	multipleAZ              bool
-	availabilityZone        *string
-	subnetId                *string
-}
-
-var lastSetup testSetup
-
-func setup1() (testSetup, context.CancelFunc) {
-	fmt.Println("********* call to before each - setting up `setup` ")
-	var err error
-	setup := testSetup{}
-	setup.testTmpDir, err = ioutil.TempDir(suiteTmpDir, "aws-test")
-	if err != nil {
-		panic("couldnt make a directory !!!!")
-	}
-	setup.namespace = "test-" + util.RandomString(6)
-	fmt.Println("*********", setup.namespace)
-	createNamespace(setup.namespace)
-
-	setup.clusterName = "test-" + util.RandomString(6)
-	setup.awsClusterName = "test-infra-" + util.RandomString(6)
-	setup.cpMachinePrefix = "test-" + util.RandomString(6)
-	setup.cpAWSMachinePrefix = "test-infra-" + util.RandomString(6)
-	setup.cpBootstrapConfigPrefix = "test-boot-" + util.RandomString(6)
-	setup.mdBootstrapConfig = "test-boot-md" + util.RandomString(6)
-	setup.machineDeploymentName = "test-capa-md" + util.RandomString(6)
-	setup.awsMachineTemplateName = "test-infra-capa-mt" + util.RandomString(6)
-	setup.initialReplicas = 2
-	setup.instanceType = "t3.large"
-	setup.multipleAZ = false
-	lastSetup = setup // so that before each has a chance in hell of watching the right setup.  not sure this works 100%.
-
-	// watch this namespace for events ...
-	ctx, cancelWatches := context.WithCancel(context.Background())
-	go func() {
-		defer GinkgoRecover()
-		watchEvents(ctx, lastSetup.namespace)
-	}()
-	return setup, cancelWatches
-}
-
 var _ = Describe("functional tests", func() {
-	var ()
+	var (
+		namespace *corev1.Namespace
+		ctx       context.Context
+	)
 
-	Describe("workload cluster lifecycle", func() {
-		It("It should be creatable and deletable", func() {
-			time.Sleep(5 * time.Second)
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a cluster with single control plane")
-			makeSingleControlPlaneCluster(setup)
-
-			By("Creating the second Control Plane Machine")
-			awsMachineName := setup.cpAWSMachinePrefix + "-1"
-			bootstrapConfigName := setup.cpBootstrapConfigPrefix + "-1"
-			machineName := setup.cpMachinePrefix + "-1"
-			createAdditionalControlPlaneMachine(setup, machineName, awsMachineName, bootstrapConfigName)
-
-			By("Creating the third Control Plane Machine")
-			awsMachineName = setup.cpAWSMachinePrefix + "-2"
-			bootstrapConfigName = setup.cpBootstrapConfigPrefix + "-2"
-			machineName = setup.cpMachinePrefix + "-2"
-			createAdditionalControlPlaneMachine(setup, machineName, awsMachineName, bootstrapConfigName)
-
-			By("Creating the MachineDeployment")
-			createMachineDeployment(setup)
-
-			By("Scale the MachineDeployment up")
-			scaleMachineDeployment(setup.namespace, setup.machineDeploymentName, setup.initialReplicas, scaleUpReplicas)
-
-			By("Scale the MachineDeployment down")
-			scaleMachineDeployment(setup.namespace, setup.machineDeploymentName, scaleUpReplicas, scaleDownReplicas)
-
-			By("Scale the MachineDeployment up again")
-			scaleMachineDeployment(setup.namespace, setup.machineDeploymentName, scaleDownReplicas, setup.initialReplicas)
-
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-		})
+	BeforeEach(func() {
+		Expect(bootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
+		ctx = context.TODO()
+		// Setup a Namespace where to host objects for this spec and create a watcher for the namespace events.
+		namespace = setupSpecNamespace(ctx, "functional-tests", bootstrapClusterProxy, artifactFolder)
 	})
 
-	Describe("Provisioning LoadBalancer dynamically and deleting on cluster deletion", func() {
-		lbServiceName := "test-svc-" + util.RandomString(6)
-		It("It should create and delete Load Balancer", func() {
-			time.Sleep(1 * time.Second)
-
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-
-			By("Creating a cluster with single control plane")
-			clusterK8sClient := makeSingleControlPlaneCluster(setup)
-
-			By("Creating the MachineDeployment")
-			createMachineDeployment(setup)
-
-			By("Creating the LB service")
-			elbName := createLBService(metav1.NamespaceDefault, lbServiceName, clusterK8sClient)
-			verifyElbExists(elbName, true)
-
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-
-			By("Verifying whether provisioned LB deleted")
-			verifyElbExists(elbName, false)
-		})
-	})
-
-	Describe("Provisioning volumes dynamically and retain on cluster deletion", func() {
+	Describe("Cluster name validations and provisioning extra AWS resources", func() {
 		nginxStatefulsetInfo := statefulSetInfo{
 			name:                      "nginx-statefulset",
 			namespace:                 metav1.NamespaceDefault,
@@ -231,645 +99,301 @@ var _ = Describe("functional tests", func() {
 			podTerminationGracePeriod: int64(30),
 			volMountPath:              "/usr/share/nginx/html",
 		}
+		It("Should create a cluster, AWS load balancer, and volume", func() {
+			By("Creating cluster with name starting with 'sg-', having '.' and more than 22 characters")
+			// Tests a cluster name that satisfies the following cases:
+			// - name with more than 22 characters
+			// - name with '.'
+			// - name that starts with 'sg-'
+			clusterName := fmt.Sprintf("sg-test.%s", util.RandomString(20))
+			configCluster := defaultConfigCluster(clusterName, namespace.Name)
+			configCluster.WorkerMachineCount = pointer.Int64Ptr(1)
+			cluster, md := createCluster(ctx, configCluster)
+			clusterClient := bootstrapClusterProxy.GetWorkloadCluster(ctx, namespace.Name, clusterName).GetClient()
 
-		It("It should create volumes and volumes should not be deleted along with cluster infra", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a cluster with single control plane")
-			clusterK8sClient := makeSingleControlPlaneCluster(setup)
+			By("Waiting for worker nodes to be in Running phase")
+			machines := getMachinesOfDeployment(namespace, md[0])
+			Expect(len(machines.Items)).Should(BeNumerically(">", 0))
+			statusChecks := []framework.MachineStatusCheck{framework.MachinePhaseCheck(string(clusterv1.MachinePhaseRunning))}
+			machineStatusInput := framework.WaitForMachineStatusCheckInput{
+				Getter:       bootstrapClusterProxy.GetClient(),
+				Machine:      &machines.Items[0],
+				StatusChecks: statusChecks,
+			}
+			framework.WaitForMachineStatusCheck(ctx, machineStatusInput, e2eConfig.GetIntervals("", "wait-machine-status")...)
 
-			By("Creating the MachineDeployment")
-			createMachineDeployment(setup)
+			By("Creating the LB service")
+			lbServiceName := "test-svc-" + util.RandomString(6)
+			elbName := createLBService(metav1.NamespaceDefault, lbServiceName, clusterClient)
+			verifyElbExists(elbName, true)
 
 			By("Deploying StatefulSet on infra")
-			createStatefulSet(nginxStatefulsetInfo, clusterK8sClient)
-			awsVolIds := getVolumeIds(nginxStatefulsetInfo, clusterK8sClient)
+			createStatefulSet(nginxStatefulsetInfo, clusterClient)
+			awsVolIds := getVolumeIds(nginxStatefulsetInfo, clusterClient)
 			verifyVolumesExists(awsVolIds)
 
 			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
+			deleteCluster(ctx, cluster)
+
+			By("Verifying whether provisioned LB deleted")
+			verifyElbExists(elbName, false)
 
 			By("Verifying dynamically provisioned volumes retention")
 			verifyVolumesExists(awsVolIds)
 
 			By("Deleting retained dynamically provisioned volumes")
 			deleteRetainedVolumes(awsVolIds)
-			By("PASSED!")
-		})
-	})
-
-	Describe("MachineDeployment with invalid subnet ID and AZ", func() {
-		It("It should be creatable and deletable", func() {
-			time.Sleep(1 * time.Second)
-
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			deployment1 := setup.machineDeploymentName + "-1"
-			deployment2 := setup.machineDeploymentName + "-2"
-			template1 := setup.awsMachineTemplateName + "-1"
-			template2 := setup.awsMachineTemplateName + "-2"
-			template3 := setup.awsMachineTemplateName + "-3"
-			bsConfig1 := setup.mdBootstrapConfig + "-1"
-			bsConfig2 := setup.mdBootstrapConfig + "-2"
-
-			By("Creating a cluster with single control plane")
-			makeSingleControlPlaneCluster(setup)
-
-			By("Creating Machine Deployment with invalid subnet ID")
-			setup.subnetId = aws.String("notcreated")
-			setup.machineDeploymentName = deployment1
-			setup.awsMachineTemplateName = template1
-			setup.mdBootstrapConfig = bsConfig1
-			createPendingMachineDeployment(setup)
-
-			By("Creating Machine Deployment in non-configured Availability Zone")
-			setup.availabilityZone = getAvailabilityZones()[1].ZoneName
-			setup.machineDeploymentName = deployment2
-			setup.awsMachineTemplateName = template2
-			setup.mdBootstrapConfig = bsConfig2
-			createPendingMachineDeployment(setup)
-
-			By("Ensuring MachineDeployments are not in running state")
-			Expect(waitForMachineDeploymentRunning(setup.namespace, deployment1)).To(BeFalse())
-			Expect(waitForMachineDeploymentRunning(setup.namespace, deployment2)).To(BeFalse())
-			eventList := getEvents(setup.namespace)
-			subnetError := "Failed to create instance: failed to run instance: InvalidSubnetID.NotFound: " +
-				"The subnet ID '%s' does not exist"
-			Expect(isErrorEventExists(setup.namespace, deployment1, "FailedCreate", fmt.Sprintf(subnetError, *setup.subnetId), eventList)).To(BeTrue())
-
-			By("Create new AwsMachineTemplate with correct Subnet ID and update its name in MachineDeployment")
-			sess = getSession()
-			clusterVpcId := getVpcId(fmt.Sprintf("%s-vpc", setup.clusterName))
-			subnetId2 := getSubnetId("vpc-id", clusterVpcId)
-			makeAWSMachineTemplate(setup.namespace, template3, setup.instanceType, nil, subnetId2)
-			setup.machineDeploymentName = deployment1
-			setup.awsMachineTemplateName = template3
-			updateMachineDeploymentInfra(setup)
-
-			By("Deleting the cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
-		})
-	})
-
-	Describe("multiple workload clusters", func() {
-		Context("in different namespaces", func() {
-			var ns1, clName1, ns2, clName2 string
-			var setup11 testSetup
-			var setup22 testSetup
-			var cf1, cf2 context.CancelFunc
-			It("should setup namespaces correctly for the two clusters...", func() {
-				time.Sleep(1 * time.Second)
-
-				setup11, cf1 = setup1()
-				setup22, cf2 = setup1()
-
-				defer cf1()
-				defer cf2()
-
-				ns1 = setup11.namespace
-				clName1 = setup11.clusterName
-
-				ns2 = setup22.namespace
-				clName2 = setup22.clusterName
-
-				By("Creating first cluster with single control plane")
-				time.Sleep(1 * time.Second)
-				makeSingleControlPlaneCluster(setup11)
-
-				By("Creating second cluster with single control plane")
-				time.Sleep(1 * time.Second)
-				makeSingleControlPlaneCluster(setup22)
-
-				By("Deleting the Clusters")
-				deleteCluster(ns1, clName1)
-				deleteCluster(ns2, clName2)
-				By("PASSED!")
-			})
-		})
-
-		Context("in same namespace", func() {
-			var setup11 testSetup
-			var setup22 testSetup
-			var cf1, cf2 context.CancelFunc
-
-			It("should create first cluster", func() {
-				setup11, cf1 = setup1()
-				setup22, cf2 = setup1()
-				defer cf1()
-				defer cf2()
-
-				By("Creating first cluster with single control plane")
-				makeSingleControlPlaneCluster(setup11)
-
-				setup22.namespace = setup11.namespace
-				By("Creating second cluster with single control plane")
-				makeSingleControlPlaneCluster(setup22)
-
-				By("Deleting the Clusters")
-				deleteCluster(setup11.namespace, setup11.clusterName)
-				deleteCluster(setup22.namespace, setup22.clusterName)
-				By("PASSED!")
-			})
-		})
-	})
-
-	Describe("MachineDeployment will replace a deleted Machine", func() {
-		It("It should reconcile the deleted machine", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane")
-			makeSingleControlPlaneCluster(setup)
-
-			By("Creating the MachineDeployment")
-			createMachineDeployment(setup)
-
-			By("Deleting a worker node machine")
-			deleteMachine(setup.namespace, setup.machineDeploymentName)
-			time.Sleep(10 * time.Second)
-
-			waitForMachineDeploymentRunning(setup.namespace, setup.machineDeploymentName)
-
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
-		})
-	})
-
-	Describe("Workload cluster in multiple AZs", func() {
-		It("It should be creatable and deletable", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane with multipleAZ=true")
-			setup.multipleAZ = true
-			makeSingleControlPlaneCluster(setup)
-
-			By("Creating the second Control Plane Machine in second AZ")
-			awsMachineName := setup.cpAWSMachinePrefix + "-1"
-			bootstrapConfigName := setup.cpBootstrapConfigPrefix + "-1"
-			machineName := setup.cpMachinePrefix + "-1"
-			setup.availabilityZone = availabilityZones[1]
-			createAdditionalControlPlaneMachine(setup, machineName, awsMachineName, bootstrapConfigName)
-
-			By("Creating the third Control Plane Machine using SubnetID in third AZ")
-			awsMachineName = setup.cpAWSMachinePrefix + "-2"
-			bootstrapConfigName = setup.cpBootstrapConfigPrefix + "-2"
-			machineName = setup.cpMachinePrefix + "-2"
-			subnetId3 := getSubnetId("cidr-block", privateCIDRs[2])
-			setup.availabilityZone = nil
-			setup.subnetId = subnetId3
-			createAdditionalControlPlaneMachine(setup, machineName, awsMachineName, bootstrapConfigName)
-
-			By("Creating the MachineDeployment in second AZ")
-			setup.availabilityZone = availabilityZones[1]
-			setup.subnetId = nil
-			createMachineDeployment(setup)
-
-			By("Creating the MachineDeployment using SubnetID in third AZ")
-			setup.awsMachineTemplateName = setup.awsMachineTemplateName + "-2"
-			setup.mdBootstrapConfig = setup.mdBootstrapConfig + "-2"
-			setup.machineDeploymentName = setup.machineDeploymentName + "-2"
-			setup.availabilityZone = nil
-			setup.subnetId = subnetId3
-			createMachineDeployment(setup)
-
-			subnetId2 := getSubnetId("cidr-block", privateCIDRs[1])
-			By("Verifying if the nodes are deployed in second AZ")
-			verfiyInstancesInSubnet(setup.initialReplicas+1, subnetId2)
-
-			By("Verifying if the nodes are deployed in third AZ")
-			verfiyInstancesInSubnet(setup.initialReplicas+1, subnetId3)
-
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
 		})
 	})
 
 	Describe("Creating cluster after reaching vpc maximum limit", func() {
 		It("Cluster created after reaching vpc limit should be in provisioning", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Create clusters till vpc limit")
-			sess = getSession()
-			limit := getElasticIPsLimit()
+			By("Create VPCs until limit is reached")
+			sess := getSession()
+			limit := getElasticIPsLimit(sess)
 			var vpcsCreated []string
-			for getCurrentVPCsCount() < limit {
-				vpcsCreated = append(vpcsCreated, createVPC("10.0.0.0/16"))
+			for getCurrentVPCsCount(sess) < limit {
+				vpcsCreated = append(vpcsCreated, createVPC(sess, "10.0.0.0/16"))
 			}
 
 			By("Creating cluster beyond vpc limit")
-			Expect(createCluster(setup.namespace, setup.clusterName, setup.awsClusterName, setup.multipleAZ)).Should(BeFalse())
-			Expect(getClusterStatus(setup.namespace, setup.clusterName)).Should(Equal(string(clusterv1.ClusterPhaseProvisioning)))
+			clusterName := fmt.Sprintf("cluster-%s", util.RandomString(6))
+			workloadClusterTemplate := clusterctl.ConfigCluster(ctx, clusterctl.ConfigClusterInput{
+				LogFolder:                filepath.Join(artifactFolder, "clusters", bootstrapClusterProxy.GetName()),
+				ClusterctlConfigPath:     clusterctlConfigPath,
+				KubeconfigPath:           bootstrapClusterProxy.GetKubeconfigPath(),
+				InfrastructureProvider:   clusterctl.DefaultInfrastructureProvider,
+				Flavor:                   clusterctl.DefaultFlavor,
+				Namespace:                namespace.Name,
+				ClusterName:              clusterName,
+				KubernetesVersion:        e2eConfig.GetVariable(KubernetesVersion),
+				ControlPlaneMachineCount: pointer.Int64Ptr(1),
+				WorkerMachineCount:       pointer.Int64Ptr(0),
+			})
+			Expect(bootstrapClusterProxy.Apply(ctx, workloadClusterTemplate)).ShouldNot(HaveOccurred())
 
 			By("Checking cluster gets provisioned when resources available")
 			if len(vpcsCreated) > 0 {
 				By("Deleting VPCs")
-				deleteVPCs(vpcsCreated)
-				By("Creating the cluster after VPCs were deleted")
-				Expect(waitForClusterInfrastructureReady(setup.namespace, setup.clusterName)).Should(BeTrue())
+				deleteVPCs(sess, vpcsCreated)
 			}
 
-			By("Deleting the cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
+			By("Waiting for cluster to reach infrastructure ready")
+			Eventually(func() bool {
+				cluster := &clusterv1.Cluster{}
+				if err := bootstrapClusterProxy.GetClient().Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace.Name, Name: clusterName}, cluster); nil == err {
+					if cluster.Status.InfrastructureReady {
+						return true
+					}
+				}
+				return false
+			}, e2eConfig.GetIntervals("", "wait-cluster")...).Should(Equal(true))
 		})
 	})
 
-	Describe("Delete infra node directly from infra provider", func() {
-		It("Machine referencing deleted infra node should come to failed state", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane")
-			makeSingleControlPlaneCluster(setup)
+	Describe("MachineDeployment misconfigurations", func() {
+		It("Should fail to create MachineDeployment with invalid subnet or non-configured Availability Zone", func() {
+			By("Creating a cluster")
+			clusterName := fmt.Sprintf("cluster-%s", util.RandomString(6))
+			configCluster := defaultConfigCluster(clusterName, namespace.Name)
+			_, _ = createCluster(ctx, configCluster)
 
-			By("Creating the MachineDeployment")
-			setup.availabilityZone = nil
-			setup.subnetId = nil
-			createMachineDeployment(setup)
+			By("Creating Machine Deployment with invalid subnet ID")
+			md1Name := clusterName + "-md-1"
+			framework.CreateMachineDeployment(ctx, framework.CreateMachineDeploymentInput{
+				Creator:                 bootstrapClusterProxy.GetClient(),
+				MachineDeployment:       makeMachineDeployment(namespace.Name, md1Name, clusterName, 1),
+				BootstrapConfigTemplate: makeJoinBootstrapConfigTemplate(namespace.Name, md1Name),
+				InfraMachineTemplate:    makeAWSMachineTemplate(namespace.Name, md1Name, e2eConfig.GetVariable(AwsNodeMachineType), nil, pointer.StringPtr("invalid-subnet")),
+			})
 
-			By("Deleting node directly from infra cloud")
-			machines, err := getMachinesOfDeployment(setup.namespace, setup.machineDeploymentName)
-			Expect(err).To(BeNil())
-			Expect(len(machines.Items)).Should(BeNumerically(">", 0))
-			terminateInstance(*machines.Items[0].Spec.ProviderID)
-			verifyMachinePhase(setup.namespace, machines.Items[0].Name, clusterv1.MachinePhaseFailed)
+			By("Looking for failure event to be reported")
+			Eventually(func() bool {
+				eventList := getEvents(namespace.Name)
+				subnetError := "Failed to create instance: failed to run instance: InvalidSubnetID.NotFound: " +
+					"The subnet ID '%s' does not exist"
+				return isErrorEventExists(namespace.Name, md1Name, "FailedCreate", fmt.Sprintf(subnetError, "invalid-subnet"), eventList)
+			}, e2eConfig.GetIntervals("", "wait-worker-nodes")...).Should(BeTrue())
 
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
+			By("Creating Machine Deployment in non-configured Availability Zone")
+			md2Name := clusterName + "-md-2"
+			//By default, first availability zone will be used for cluster resources. This step attempts to create a machine deployment in the second availability zone
+			invalidAz := getAvailabilityZones()[1].ZoneName
+			framework.CreateMachineDeployment(ctx, framework.CreateMachineDeploymentInput{
+				Creator:                 bootstrapClusterProxy.GetClient(),
+				MachineDeployment:       makeMachineDeployment(namespace.Name, md2Name, clusterName, 1),
+				BootstrapConfigTemplate: makeJoinBootstrapConfigTemplate(namespace.Name, md2Name),
+				InfraMachineTemplate:    makeAWSMachineTemplate(namespace.Name, md2Name, e2eConfig.GetVariable(AwsNodeMachineType), invalidAz, nil),
+			})
+
+			By("Looking for failure event to be reported")
+			Eventually(func() bool {
+				eventList := getEvents(namespace.Name)
+				azError := "Failed to create instance: no subnets available in availability zone \"%s\""
+				return isErrorEventExists(namespace.Name, md2Name, "FailedCreate", fmt.Sprintf(azError, *invalidAz), eventList)
+			}, e2eConfig.GetIntervals("", "wait-worker-nodes")...).Should(BeTrue())
 		})
 	})
 
-	Describe("Create cluster with name having more than 22 characters", func() {
-		It("Cluster should be provisioned and deleted", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane")
-			setup.clusterName = "test-" + util.RandomString(20)
-			makeSingleControlPlaneCluster(setup)
+	Describe("Workload cluster in multiple AZs", func() {
+		It("It should be creatable and deletable", func() {
+			By("Creating a cluster")
+			clusterName := fmt.Sprintf("cluster-%s", util.RandomString(6))
+			configCluster := defaultConfigCluster(clusterName, namespace.Name)
+			configCluster.ControlPlaneMachineCount = pointer.Int64Ptr(3)
+			configCluster.Flavor = MultiAzFlavor
+			cluster, _ := createCluster(ctx, configCluster)
 
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
+			By("Adding worker nodes to additional subnets")
+			mdName1 := clusterName + "-md-1"
+			mdName2 := clusterName + "-md-2"
+			md1 := makeMachineDeployment(namespace.Name, mdName1, clusterName, 1)
+			md2 := makeMachineDeployment(namespace.Name, mdName2, clusterName, 1)
+			az1 := os.Getenv(AwsAvailabilityZone1)
+			az2 := os.Getenv(AwsAvailabilityZone2)
+
+			//private CIDRs set in cluster-template-multi-az.yaml
+			framework.CreateMachineDeployment(ctx, framework.CreateMachineDeploymentInput{
+				Creator:                 bootstrapClusterProxy.GetClient(),
+				MachineDeployment:       md1,
+				BootstrapConfigTemplate: makeJoinBootstrapConfigTemplate(namespace.Name, mdName1),
+				InfraMachineTemplate:    makeAWSMachineTemplate(namespace.Name, mdName1, e2eConfig.GetVariable(AwsNodeMachineType), pointer.StringPtr(az1), getSubnetId("cidr-block", "10.0.0.0/24")),
+			})
+			framework.CreateMachineDeployment(ctx, framework.CreateMachineDeploymentInput{
+				Creator:                 bootstrapClusterProxy.GetClient(),
+				MachineDeployment:       md2,
+				BootstrapConfigTemplate: makeJoinBootstrapConfigTemplate(namespace.Name, mdName2),
+				InfraMachineTemplate:    makeAWSMachineTemplate(namespace.Name, mdName2, e2eConfig.GetVariable(AwsNodeMachineType), pointer.StringPtr(az2), getSubnetId("cidr-block", "10.0.2.0/24")),
+			})
+
+			By("Waiting for new worker nodes to become ready")
+			k8sClient := bootstrapClusterProxy.GetClient()
+			framework.WaitForMachineDeploymentNodesToExist(ctx, framework.WaitForMachineDeploymentNodesToExistInput{Lister: k8sClient, Cluster: cluster, MachineDeployment: md1}, e2eConfig.GetIntervals("", "wait-worker-nodes")...)
+			framework.WaitForMachineDeploymentNodesToExist(ctx, framework.WaitForMachineDeploymentNodesToExistInput{Lister: k8sClient, Cluster: cluster, MachineDeployment: md2}, e2eConfig.GetIntervals("", "wait-worker-nodes")...)
 		})
 	})
 
-	Describe("Create cluster with name having '.'", func() {
-		It("Cluster should be provisioned and deleted", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane")
-			setup.clusterName = "test." + util.RandomString(6)
-			makeSingleControlPlaneCluster(setup)
+	Describe("multiple workload clusters", func() {
+		Context("in different namespaces with machine failures", func() {
+			It("should setup namespaces correctly for the two clusters", func() {
+				By("Creating first cluster with single control plane")
+				ns1, cf1 := framework.CreateNamespaceAndWatchEvents(ctx, framework.CreateNamespaceAndWatchEventsInput{
+					Creator:   bootstrapClusterProxy.GetClient(),
+					ClientSet: bootstrapClusterProxy.GetClientSet(),
+					Name:      fmt.Sprintf("multi-workload-%s", util.RandomString(6)),
+					LogFolder: filepath.Join(artifactFolder, "clusters", bootstrapClusterProxy.GetName()),
+				})
+				ns2, cf2 := framework.CreateNamespaceAndWatchEvents(ctx, framework.CreateNamespaceAndWatchEventsInput{
+					Creator:   bootstrapClusterProxy.GetClient(),
+					ClientSet: bootstrapClusterProxy.GetClientSet(),
+					Name:      fmt.Sprintf("multi-workload-%s", util.RandomString(6)),
+					LogFolder: filepath.Join(artifactFolder, "clusters", bootstrapClusterProxy.GetName()),
+				})
 
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
+				By("Creating first cluster")
+				cluster1Name := fmt.Sprintf("cluster-%s", util.RandomString(6))
+				configCluster := defaultConfigCluster(cluster1Name, ns1.Name)
+				configCluster.WorkerMachineCount = pointer.Int64Ptr(1)
+				configCluster.Flavor = LimitAzFlavor
+				cluster1, md1 := createCluster(ctx, configCluster)
+				Expect(len(md1)).To(Equal(1), "Expecting one MachineDeployment")
+
+				By("Deleting a worker node machine")
+				deleteMachine(ns1, md1[0])
+				time.Sleep(10 * time.Second)
+
+				By("Verifying MachineDeployment is running.")
+				framework.DiscoveryAndWaitForMachineDeployments(ctx, framework.DiscoveryAndWaitForMachineDeploymentsInput{Cluster: cluster1, Lister: bootstrapClusterProxy.GetClient()}, e2eConfig.GetIntervals("", "wait-worker-nodes")...)
+
+				By("Creating second cluster")
+				cluster2Name := fmt.Sprintf("cluster-%s", util.RandomString(6))
+				configCluster = defaultConfigCluster(cluster2Name, ns2.Name)
+				configCluster.WorkerMachineCount = pointer.Int64Ptr(1)
+				configCluster.Flavor = LimitAzFlavor
+				cluster2, md2 := createCluster(ctx, configCluster)
+				Expect(len(md2)).To(Equal(1), "Expecting one MachineDeployment")
+
+				By("Deleting node directly from infra cloud")
+				machines := getMachinesOfDeployment(ns2, md2[0])
+				Expect(len(machines.Items)).Should(BeNumerically(">", 0))
+				terminateInstance(*machines.Items[0].Spec.ProviderID)
+
+				By("Waiting for machine to reach Failed state")
+				statusChecks := []framework.MachineStatusCheck{framework.MachinePhaseCheck(string(clusterv1.MachinePhaseFailed))}
+				machineStatusInput := framework.WaitForMachineStatusCheckInput{
+					Getter:       bootstrapClusterProxy.GetClient(),
+					Machine:      &machines.Items[0],
+					StatusChecks: statusChecks,
+				}
+				framework.WaitForMachineStatusCheck(ctx, machineStatusInput, e2eConfig.GetIntervals("", "wait-machine-status")...)
+
+				By("Deleting the clusters and namespaces")
+				deleteCluster(ctx, cluster1)
+				deleteCluster(ctx, cluster2)
+				framework.DeleteNamespace(ctx, framework.DeleteNamespaceInput{Deleter: bootstrapClusterProxy.GetClient(), Name: ns1.Name})
+				framework.DeleteNamespace(ctx, framework.DeleteNamespaceInput{Deleter: bootstrapClusterProxy.GetClient(), Name: ns2.Name})
+				cf1()
+				cf2()
+			})
+		})
+
+		Context("in same namespace", func() {
+			It("should create the clusters", func() {
+				By("Creating first cluster with single control plane")
+				cluster1Name := fmt.Sprintf("cluster-%s", util.RandomString(6))
+				configCluster := defaultConfigCluster(cluster1Name, namespace.Name)
+				configCluster.Flavor = LimitAzFlavor
+				cluster1, _ := createCluster(ctx, configCluster)
+
+				By("Creating second cluster with single control plane")
+				cluster2Name := fmt.Sprintf("cluster-%s", util.RandomString(6))
+				configCluster = defaultConfigCluster(cluster2Name, namespace.Name)
+				configCluster.Flavor = LimitAzFlavor
+				cluster2, _ := createCluster(ctx, configCluster)
+
+				By("Deleting the Clusters")
+				deleteCluster(ctx, cluster1)
+				deleteCluster(ctx, cluster2)
+			})
 		})
 	})
 
-	Describe("Create cluster with name starting with 'sg-'", func() {
-		It("Cluster should be provisioned and deleted", func() {
-			setup, cancelFunc := setup1()
-			defer cancelFunc()
-			By("Creating a workload cluster with single control plane")
-			setup.clusterName = "sg-" + util.RandomString(6)
-			makeSingleControlPlaneCluster(setup)
-
-			By("Deleting the Cluster")
-			deleteCluster(setup.namespace, setup.clusterName)
-			By("PASSED!")
-		})
+	AfterEach(func() {
+		// Dumps all the resources in the spec namespace, then cleanups the cluster object and the spec namespace itself.
+		dumpSpecResourcesAndCleanup(ctx, "", bootstrapClusterProxy, artifactFolder, namespace, e2eConfig.GetIntervals, skipCleanup)
 	})
 })
 
-func watchEvents(ctx context.Context, namespace string) {
-	logFile := path.Join(artifactPath, "resources", namespace, "events.log")
-	fmt.Fprintf(GinkgoWriter, "Creating directory: %s\n", filepath.Dir(logFile))
-	Expect(os.MkdirAll(filepath.Dir(logFile), 0o755)).To(Succeed())
-
-	f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644)
-	Expect(err).NotTo(HaveOccurred())
-	defer f.Close()
-
-	informerFactory := informers.NewSharedInformerFactoryWithOptions(
-		clientSet,
-		10*time.Minute,
-		informers.WithNamespace(namespace),
-	)
-	eventInformer := informerFactory.Core().V1().Events().Informer()
-	eventInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			e := obj.(*corev1.Event)
-			f.WriteString(fmt.Sprintf("[New Event] %s/%s\n\tresource: %s/%s/%s\n\treason: %s\n\tmessage: %s\n\tfull: %#v\n",
-				e.Namespace, e.Name, e.InvolvedObject.APIVersion, e.InvolvedObject.Kind, e.InvolvedObject.Name, e.Reason, e.Message, e))
-		},
-		UpdateFunc: func(_, obj interface{}) {
-			e := obj.(*corev1.Event)
-			f.WriteString(fmt.Sprintf("[Updated Event] %s/%s\n\tresource: %s/%s/%s\n\treason: %s\n\tmessage: %s\n\tfull: %#v\n",
-				e.Namespace, e.Name, e.InvolvedObject.APIVersion, e.InvolvedObject.Kind, e.InvolvedObject.Name, e.Reason, e.Message, e))
-		},
-		DeleteFunc: func(obj interface{}) {},
+func createCluster(ctx context.Context, configCluster clusterctl.ConfigClusterInput) (*clusterv1.Cluster, []*clusterv1.MachineDeployment) {
+	cluster, _, md := clusterctl.ApplyClusterTemplateAndWait(ctx, clusterctl.ApplyClusterTemplateAndWaitInput{
+		ClusterProxy:                 bootstrapClusterProxy,
+		ConfigCluster:                configCluster,
+		CNIManifestPath:              e2eConfig.GetVariable(CNIPath),
+		WaitForClusterIntervals:      e2eConfig.GetIntervals("", "wait-cluster"),
+		WaitForControlPlaneIntervals: e2eConfig.GetIntervals("", "wait-control-plane"),
+		WaitForMachineDeployments:    e2eConfig.GetIntervals("", "wait-worker-nodes"),
 	})
 
-	stopInformer := make(chan struct{})
-	defer close(stopInformer)
-	informerFactory.Start(stopInformer)
-	<-ctx.Done()
-	stopInformer <- struct{}{}
+	return cluster, md
 }
 
-func updateMachineDeploymentInfra(setup testSetup) {
-	deployment := &clusterv1.MachineDeployment{}
-	Expect(kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: setup.namespace, Name: setup.machineDeploymentName}, deployment))
-	deployment.Spec.Template.Spec.InfrastructureRef.Name = setup.awsMachineTemplateName
-	Expect(kindClient.Update(context.TODO(), deployment)).NotTo(HaveOccurred())
-	waitForMachinesCountMatch(setup.namespace, setup.machineDeploymentName, setup.initialReplicas, setup.initialReplicas)
-	waitForMachineDeploymentRunning(setup.namespace, setup.machineDeploymentName)
-}
-
-func verifyMachinePhase(namespace, machineName string, phase clusterv1.MachinePhase) {
-	fmt.Fprintf(GinkgoWriter, "Ensuring machine %s's state is %s ... \n", machineName, string(phase))
-	machine := &clusterv1.Machine{}
-	Eventually(
-		func() bool {
-			if err := kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineName}, machine); err != nil {
-				return false
-			}
-			return machine.Status.Phase == string(phase)
-		}, 20*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func terminateInstance(instanceId string) {
-	fmt.Fprintf(GinkgoWriter, "Terminating EC2 instance with ID: %s ... \n", instanceId)
-	ec2Client := ec2.New(getSession())
-	input := &ec2.TerminateInstancesInput{
-		InstanceIds: []*string{
-			aws.String(instanceId[strings.LastIndex(instanceId, "/")+1:]),
-		},
+func defaultConfigCluster(clusterName, namespace string) clusterctl.ConfigClusterInput {
+	return clusterctl.ConfigClusterInput{
+		LogFolder:                filepath.Join(artifactFolder, "clusters", bootstrapClusterProxy.GetName()),
+		ClusterctlConfigPath:     clusterctlConfigPath,
+		KubeconfigPath:           bootstrapClusterProxy.GetKubeconfigPath(),
+		InfrastructureProvider:   clusterctl.DefaultInfrastructureProvider,
+		Flavor:                   clusterctl.DefaultFlavor,
+		Namespace:                namespace,
+		ClusterName:              clusterName,
+		KubernetesVersion:        e2eConfig.GetVariable(KubernetesVersion),
+		ControlPlaneMachineCount: pointer.Int64Ptr(1),
+		WorkerMachineCount:       pointer.Int64Ptr(0),
 	}
-
-	result, err := ec2Client.TerminateInstances(input)
-	Expect(err).To(BeNil())
-	Expect(len(result.TerminatingInstances)).To(Equal(1))
-	termCode := int64(32)
-	Expect(*result.TerminatingInstances[0].CurrentState.Code).To(Equal(termCode))
-}
-
-func getMachinesOfDeployment(namespace, machineDeploymentName string) (*clusterv1.MachineList, error) {
-	fmt.Fprintf(GinkgoWriter, "Fetching the Machines of MachineDeployment %s: \n", machineDeploymentName)
-	machineList := &clusterv1.MachineList{}
-	machineDeployment := &clusterv1.MachineDeployment{}
-	if err := kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineDeploymentName}, machineDeployment); err != nil {
-		fmt.Fprintf(GinkgoWriter, "Got error while getting machinedeployment %s \n", machineDeploymentName)
-		return nil, err
-	}
-
-	selector, err := metav1.LabelSelectorAsMap(&machineDeployment.Spec.Selector)
-	if err != nil {
-		fmt.Fprintf(GinkgoWriter, "Got error while reading lables of machinedeployment: %s, %s \n", machineDeploymentName, err.Error())
-		return nil, err
-	}
-
-	if err := kindClient.List(context.TODO(), machineList, crclient.InNamespace(namespace), crclient.MatchingLabels(selector)); err != nil {
-		fmt.Fprintf(GinkgoWriter, "Got error while getting machines of machinedeployment: %s, %s \n", machineDeploymentName, err.Error())
-		return nil, err
-	}
-	return machineList, nil
-}
-
-func deleteVPCs(vpcIds []string) {
-	ec2Client := ec2.New(sess)
-	for _, vpcId := range vpcIds {
-		input := &ec2.DeleteVpcInput{
-			VpcId: aws.String(vpcId),
-		}
-		_, err := ec2Client.DeleteVpc(input)
-		Expect(err).NotTo(HaveOccurred())
-	}
-}
-
-func createVPC(cidrblock string) string {
-	ec2Client := ec2.New(sess)
-	input := &ec2.CreateVpcInput{
-		CidrBlock: aws.String(cidrblock),
-	}
-	result, err := ec2Client.CreateVpc(input)
-	Expect(err).NotTo(HaveOccurred())
-	return *result.Vpc.VpcId
-}
-
-func getClusterStatus(namespace, clusterName string) string {
-	cluster := &clusterv1.Cluster{}
-	if err := kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: clusterName}, cluster); nil == err {
-		return cluster.Status.Phase
-	}
-	return ""
-}
-
-func getElasticIPsLimit() int {
-	ec2Client := ec2.New(sess)
-	input := &ec2.DescribeAccountAttributesInput{
-		AttributeNames: []*string{
-			aws.String("vpc-max-elastic-ips"),
-		},
-	}
-	result, err := ec2Client.DescribeAccountAttributes(input)
-	Expect(err).NotTo(HaveOccurred())
-	res, err := strconv.Atoi(*result.AccountAttributes[0].AttributeValues[0].AttributeValue)
-	Expect(err).NotTo(HaveOccurred())
-	return res
-}
-
-func getCurrentVPCsCount() int {
-	ec2Client := ec2.New(sess)
-	input := &ec2.DescribeVpcsInput{}
-	result, err := ec2Client.DescribeVpcs(input)
-	Expect(err).NotTo(HaveOccurred())
-	return len(result.Vpcs)
-}
-
-func getAvailabilityZones() []*ec2.AvailabilityZone {
-	ec2Client := ec2.New(getSession())
-	azs, err := ec2Client.DescribeAvailabilityZones(nil)
-	Expect(err).NotTo(HaveOccurred())
-	return azs.AvailabilityZones
-}
-
-func createStatefulSet(statefulsetinfo statefulSetInfo, k8sclient crclient.Client) {
-	fmt.Fprintf(GinkgoWriter, "Creating statefulset...\n")
-	createStorageClass(statefulsetinfo.storageClassName, k8sclient)
-	svcSpec := corev1.ServiceSpec{
-		ClusterIP: "None",
-		Ports: []corev1.ServicePort{
-			{
-				Port: statefulsetinfo.svcPort,
-				Name: statefulsetinfo.svcPortName,
-			},
-		},
-		Selector: statefulsetinfo.selector,
-	}
-	createService(statefulsetinfo.svcName, statefulsetinfo.namespace, statefulsetinfo.selector, svcSpec, k8sclient)
-	podTemplateSpec := createPodTemplateSpec(statefulsetinfo)
-	volClaimTemplate := createPVC(statefulsetinfo)
-	deployStatefulSet(statefulsetinfo, volClaimTemplate, podTemplateSpec, k8sclient)
-	waitForStatefulSetRunning(statefulsetinfo, k8sclient)
-}
-
-func deleteRetainedVolumes(awsVolIds []*string) {
-	fmt.Fprintf(GinkgoWriter, "Deleting dynamically provisioned volumes...\n")
-	ec2Client := ec2.New(getSession())
-	for _, volumeId := range awsVolIds {
-		input := &ec2.DeleteVolumeInput{
-			VolumeId: aws.String(*volumeId),
-		}
-		_, err := ec2Client.DeleteVolume(input)
-		Expect(err).NotTo(HaveOccurred())
-		fmt.Fprintf(GinkgoWriter, "Deleted dynamically provisioned volume with ID: %s \n", *volumeId)
-	}
-}
-
-func verifyVolumesExists(awsVolumeIds []*string) {
-	fmt.Fprintf(GinkgoWriter, "Ensuring dynamically provisioned volumes exists..\n")
-	ec2Client := ec2.New(getSession())
-	input := &ec2.DescribeVolumesInput{
-		VolumeIds: awsVolumeIds,
-	}
-	_, err := ec2Client.DescribeVolumes(input)
-	Expect(err).NotTo(HaveOccurred())
-}
-
-func getVolumeIds(info statefulSetInfo, k8sclient crclient.Client) []*string {
-	fmt.Fprintf(GinkgoWriter, "Retrieving IDs of dynamically provisioned volumes..\n")
-	statefulset := &appsv1.StatefulSet{}
-	err := k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: info.name}, statefulset)
-	Expect(err).NotTo(HaveOccurred())
-	podSelector, err := metav1.LabelSelectorAsMap(statefulset.Spec.Selector)
-	pvcList := &corev1.PersistentVolumeClaimList{}
-	err = k8sclient.List(context.TODO(), pvcList, crclient.InNamespace(info.namespace), crclient.MatchingLabels(podSelector))
-	Expect(err).NotTo(HaveOccurred())
-	var volIds []*string
-	for _, pvc := range pvcList.Items {
-		volName := pvc.Spec.VolumeName
-		volDescription := &corev1.PersistentVolume{}
-		err = k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: volName}, volDescription)
-		Expect(err).NotTo(HaveOccurred())
-		urlSlice := strings.Split(volDescription.Spec.AWSElasticBlockStore.VolumeID, "/")
-		volIds = append(volIds, &urlSlice[len(urlSlice)-1])
-	}
-	return volIds
-}
-
-func waitForStatefulSetRunning(info statefulSetInfo, k8sclient crclient.Client) {
-	fmt.Fprintf(GinkgoWriter, "Ensuring Statefulset(%s) is running..\n", info.name)
-	statefulset := &appsv1.StatefulSet{}
-	Eventually(
-		func() (bool, error) {
-			if err := k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: info.name}, statefulset); err != nil {
-				return false, err
-			}
-			return *statefulset.Spec.Replicas == statefulset.Status.ReadyReplicas, nil
-		}, 10*time.Minute, 30*time.Second,
-	).Should(BeTrue())
-}
-
-func deployStatefulSet(statefulsetinfo statefulSetInfo, volClaimTemp corev1.PersistentVolumeClaim, podTemplate corev1.PodTemplateSpec, k8sclient crclient.Client) {
-	fmt.Fprintf(GinkgoWriter, "Deploying Statefulset with name: %s under namespace: %s\n", statefulsetinfo.name, statefulsetinfo.namespace)
-	statefulset := appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{Name: statefulsetinfo.name, Namespace: statefulsetinfo.namespace},
-		Spec: appsv1.StatefulSetSpec{
-			Replicas:             &statefulsetinfo.replicas,
-			Selector:             &metav1.LabelSelector{MatchLabels: statefulsetinfo.selector},
-			Template:             podTemplate,
-			VolumeClaimTemplates: []corev1.PersistentVolumeClaim{volClaimTemp},
-		},
-	}
-	Expect(k8sclient.Create(context.TODO(), &statefulset)).NotTo(HaveOccurred())
-}
-
-func createPVC(statefulsetinfo statefulSetInfo) corev1.PersistentVolumeClaim {
-	fmt.Fprintf(GinkgoWriter, "Creating PersistentVolumeClaim config object\n")
-	volClaimTemplate := corev1.PersistentVolumeClaim{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: statefulsetinfo.volumeName,
-		},
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes:      []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-			StorageClassName: &statefulsetinfo.storageClassName,
-			Resources: corev1.ResourceRequirements{
-				Requests: corev1.ResourceList{
-					corev1.ResourceName(corev1.ResourceStorage): resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-	return volClaimTemplate
-}
-
-func createPodTemplateSpec(statefulsetinfo statefulSetInfo) corev1.PodTemplateSpec {
-	fmt.Fprintf(GinkgoWriter, "Creating PodTemplateSpec config object\n")
-	podTemplateSpec := corev1.PodTemplateSpec{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels: statefulsetinfo.selector,
-		},
-		Spec: corev1.PodSpec{
-			TerminationGracePeriodSeconds: &statefulsetinfo.podTerminationGracePeriod,
-			Containers: []corev1.Container{
-				{
-					Name:  statefulsetinfo.containerName,
-					Image: statefulsetinfo.containerImage,
-					Ports: []corev1.ContainerPort{{Name: statefulsetinfo.svcPortName, ContainerPort: statefulsetinfo.containerPort}},
-					VolumeMounts: []corev1.VolumeMount{
-						{Name: statefulsetinfo.volumeName, MountPath: statefulsetinfo.volMountPath},
-					},
-				},
-			},
-		},
-	}
-	return podTemplateSpec
-}
-
-func createStorageClass(storageClassName string, k8sclient crclient.Client) {
-	fmt.Fprintf(GinkgoWriter, "Creating StorageClass object with name: %s\n", storageClassName)
-	volExpansion := true
-	bindingMode := storagev1.VolumeBindingImmediate
-	storageClass := storagev1.StorageClass{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "storage.k8s.io/v1",
-			Kind:       "StorageClass",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: storageClassName,
-		},
-		Parameters: map[string]string{
-			"type": "gp2",
-		},
-		Provisioner:          "kubernetes.io/aws-ebs",
-		AllowVolumeExpansion: &volExpansion,
-		MountOptions:         []string{"debug"},
-		VolumeBindingMode:    &bindingMode,
-	}
-	Expect(k8sclient.Create(context.TODO(), &storageClass)).NotTo(HaveOccurred())
-}
-
-func createService(svcName string, svcNamespace string, labels map[string]string, serviceSpec corev1.ServiceSpec, k8sClient crclient.Client) {
-	svcToCreate := corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: svcNamespace,
-			Name:      svcName,
-		},
-		Spec: serviceSpec,
-	}
-	if labels != nil && len(labels) > 0 {
-		svcToCreate.ObjectMeta.Labels = labels
-	}
-	Expect(k8sClient.Create(context.TODO(), &svcToCreate)).NotTo(HaveOccurred())
 }
 
 func createLBService(svcNamespace string, svcName string, k8sclient crclient.Client) string {
-	fmt.Fprintf(GinkgoWriter, "Creating service of type Load Balancer with name: %s under namespace: %s\n", svcName, svcNamespace)
+	Byf("Creating service of type Load Balancer with name: %s under namespace: %s", svcName, svcNamespace)
 	svcSpec := corev1.ServiceSpec{
 		Type: corev1.ServiceTypeLoadBalancer,
 		Ports: []corev1.ServicePort{
@@ -893,56 +417,260 @@ func createLBService(svcNamespace string, svcName string, k8sclient crclient.Cli
 		ingressHostname := svcCreated.Status.LoadBalancer.Ingress[0].Hostname
 		elbName = strings.Split(ingressHostname, "-")[0]
 	}
-	fmt.Fprintf(GinkgoWriter, "Created Load Balancer service and ELB name is: %s\n", elbName)
+	Byf("Created Load Balancer service and ELB name is: %s", elbName)
+
 	return elbName
 }
 
-func verifyElbExists(elbName string, exists bool) {
-	fmt.Fprintf(GinkgoWriter, "Verifying ELB with name %s present\n", elbName)
-	elbClient := elb.New(getSession())
-	input := &elb.DescribeLoadBalancersInput{
-		LoadBalancerNames: []*string{
-			aws.String(elbName),
+func createPodTemplateSpec(statefulsetinfo statefulSetInfo) corev1.PodTemplateSpec {
+	By("Creating PodTemplateSpec config object")
+	podTemplateSpec := corev1.PodTemplateSpec{
+		ObjectMeta: metav1.ObjectMeta{
+			Labels: statefulsetinfo.selector,
+		},
+		Spec: corev1.PodSpec{
+			TerminationGracePeriodSeconds: &statefulsetinfo.podTerminationGracePeriod,
+			Containers: []corev1.Container{
+				{
+					Name:  statefulsetinfo.containerName,
+					Image: statefulsetinfo.containerImage,
+					Ports: []corev1.ContainerPort{{Name: statefulsetinfo.svcPortName, ContainerPort: statefulsetinfo.containerPort}},
+					VolumeMounts: []corev1.VolumeMount{
+						{Name: statefulsetinfo.volumeName, MountPath: statefulsetinfo.volMountPath},
+					},
+				},
+			},
 		},
 	}
-	elbsOutput, err := elbClient.DescribeLoadBalancers(input)
-	if exists {
+	return podTemplateSpec
+}
+
+func createPVC(statefulsetinfo statefulSetInfo) corev1.PersistentVolumeClaim {
+	By("Creating PersistentVolumeClaim config object")
+	volClaimTemplate := corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: statefulsetinfo.volumeName,
+		},
+		Spec: corev1.PersistentVolumeClaimSpec{
+			AccessModes:      []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
+			StorageClassName: &statefulsetinfo.storageClassName,
+			Resources: corev1.ResourceRequirements{
+				Requests: corev1.ResourceList{
+					corev1.ResourceName(corev1.ResourceStorage): resource.MustParse("1Gi"),
+				},
+			},
+		},
+	}
+	return volClaimTemplate
+}
+
+func createService(svcName string, svcNamespace string, labels map[string]string, serviceSpec corev1.ServiceSpec, k8sClient crclient.Client) {
+	svcToCreate := corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: svcNamespace,
+			Name:      svcName,
+		},
+		Spec: serviceSpec,
+	}
+	if len(labels) > 0 {
+		svcToCreate.ObjectMeta.Labels = labels
+	}
+	Expect(k8sClient.Create(context.TODO(), &svcToCreate)).NotTo(HaveOccurred())
+}
+
+func createStatefulSet(statefulsetinfo statefulSetInfo, k8sclient crclient.Client) {
+	By("Creating statefulset")
+	createStorageClass(statefulsetinfo.storageClassName, k8sclient)
+	svcSpec := corev1.ServiceSpec{
+		ClusterIP: "None",
+		Ports: []corev1.ServicePort{
+			{
+				Port: statefulsetinfo.svcPort,
+				Name: statefulsetinfo.svcPortName,
+			},
+		},
+		Selector: statefulsetinfo.selector,
+	}
+	createService(statefulsetinfo.svcName, statefulsetinfo.namespace, statefulsetinfo.selector, svcSpec, k8sclient)
+	podTemplateSpec := createPodTemplateSpec(statefulsetinfo)
+	volClaimTemplate := createPVC(statefulsetinfo)
+	deployStatefulSet(statefulsetinfo, volClaimTemplate, podTemplateSpec, k8sclient)
+	waitForStatefulSetRunning(statefulsetinfo, k8sclient)
+}
+
+func createStorageClass(storageClassName string, k8sclient crclient.Client) {
+	Byf("Creating StorageClass object with name: %s", storageClassName)
+	volExpansion := true
+	bindingMode := storagev1.VolumeBindingImmediate
+	azs := getAvailabilityZones()
+	storageClass := storagev1.StorageClass{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "storage.k8s.io/v1",
+			Kind:       "StorageClass",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: storageClassName,
+		},
+		Parameters: map[string]string{
+			"type": "gp2",
+		},
+		Provisioner:          "kubernetes.io/aws-ebs",
+		AllowVolumeExpansion: &volExpansion,
+		MountOptions:         []string{"debug"},
+		VolumeBindingMode:    &bindingMode,
+		AllowedTopologies: []corev1.TopologySelectorTerm{{
+			MatchLabelExpressions: []corev1.TopologySelectorLabelRequirement{{
+				Key:    StorageClassFailureZoneLabel,
+				Values: []string{*azs[0].ZoneName},
+			}},
+		}},
+	}
+	Expect(k8sclient.Create(context.TODO(), &storageClass)).NotTo(HaveOccurred())
+}
+
+func createVPC(sess client.ConfigProvider, cidrblock string) string {
+	ec2Client := ec2.New(sess)
+	input := &ec2.CreateVpcInput{
+		CidrBlock: aws.String(cidrblock),
+	}
+	result, err := ec2Client.CreateVpc(input)
+	Expect(err).NotTo(HaveOccurred())
+	return *result.Vpc.VpcId
+}
+
+func deleteCluster(ctx context.Context, cluster *clusterv1.Cluster) {
+	framework.DeleteCluster(ctx, framework.DeleteClusterInput{
+		Deleter: bootstrapClusterProxy.GetClient(),
+		Cluster: cluster,
+	})
+
+	framework.WaitForClusterDeleted(ctx, framework.WaitForClusterDeletedInput{
+		Getter:  bootstrapClusterProxy.GetClient(),
+		Cluster: cluster,
+	}, e2eConfig.GetIntervals("", "wait-delete-cluster")...)
+}
+
+func deleteMachine(namespace *corev1.Namespace, md *clusterv1.MachineDeployment) {
+	machineList := &clusterv1.MachineList{}
+	selector, err := metav1.LabelSelectorAsMap(&md.Spec.Selector)
+	Expect(err).NotTo(HaveOccurred())
+
+	bootstrapClient := bootstrapClusterProxy.GetClient()
+	err = bootstrapClient.List(context.TODO(), machineList, crclient.InNamespace(namespace.Name), crclient.MatchingLabels(selector))
+	Expect(err).NotTo(HaveOccurred())
+
+	Expect(len(machineList.Items)).ToNot(Equal(0))
+	machine := &clusterv1.Machine{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: namespace.Name,
+			Name:      machineList.Items[0].Name,
+		},
+	}
+	Expect(bootstrapClient.Delete(context.TODO(), machine)).To(Succeed())
+}
+
+func deleteRetainedVolumes(awsVolIds []*string) {
+	By("Deleting dynamically provisioned volumes")
+	ec2Client := ec2.New(getSession())
+	for _, volumeId := range awsVolIds {
+		input := &ec2.DeleteVolumeInput{
+			VolumeId: aws.String(*volumeId),
+		}
+		_, err := ec2Client.DeleteVolume(input)
 		Expect(err).NotTo(HaveOccurred())
-		Expect(len(elbsOutput.LoadBalancerDescriptions)).To(Equal(1))
-		fmt.Fprintf(GinkgoWriter, "ELB with name %s exists\n", elbName)
-	} else {
-		aerr, ok := err.(awserr.Error)
-		Expect(ok).To(BeTrue())
-		Expect(aerr.Code()).To(Equal(elb.ErrCodeAccessPointNotFoundException))
-		fmt.Fprintf(GinkgoWriter, "ELB with name %s doesn't exists\n", elbName)
+		Byf("Deleted dynamically provisioned volume with ID: %s", *volumeId)
 	}
 }
 
-func createClusterKubeConfigs(tmpDir, namespace, clusterName string) (string, crclient.Client) {
-	cluster := crclient.ObjectKey{
-		Namespace: namespace,
-		Name:      clusterName,
+func deleteVPCs(sess client.ConfigProvider, vpcIds []string) {
+	ec2Client := ec2.New(sess)
+	for _, vpcId := range vpcIds {
+		input := &ec2.DeleteVpcInput{
+			VpcId: aws.String(vpcId),
+		}
+		_, err := ec2Client.DeleteVpc(input)
+		Expect(err).NotTo(HaveOccurred())
 	}
-	kubeConfigData, err := kubeconfig.FromSecret(context.TODO(), kindClient, cluster)
+}
+
+func deployStatefulSet(statefulsetinfo statefulSetInfo, volClaimTemp corev1.PersistentVolumeClaim, podTemplate corev1.PodTemplateSpec, k8sclient crclient.Client) {
+	Byf("Deploying Statefulset with name: %s under namespace: %s", statefulsetinfo.name, statefulsetinfo.namespace)
+	statefulset := appsv1.StatefulSet{
+		ObjectMeta: metav1.ObjectMeta{Name: statefulsetinfo.name, Namespace: statefulsetinfo.namespace},
+		Spec: appsv1.StatefulSetSpec{
+			Replicas:             &statefulsetinfo.replicas,
+			Selector:             &metav1.LabelSelector{MatchLabels: statefulsetinfo.selector},
+			Template:             podTemplate,
+			VolumeClaimTemplates: []corev1.PersistentVolumeClaim{volClaimTemp},
+		},
+	}
+	Expect(k8sclient.Create(context.TODO(), &statefulset)).NotTo(HaveOccurred())
+}
+
+func getAvailabilityZones() []*ec2.AvailabilityZone {
+	ec2Client := ec2.New(getSession())
+	azs, err := ec2Client.DescribeAvailabilityZones(nil)
+	Expect(err).NotTo(HaveOccurred())
+	return azs.AvailabilityZones
+}
+
+func getCurrentVPCsCount(sess client.ConfigProvider) int {
+	ec2Client := ec2.New(sess)
+	input := &ec2.DescribeVpcsInput{}
+	result, err := ec2Client.DescribeVpcs(input)
+	Expect(err).NotTo(HaveOccurred())
+	return len(result.Vpcs)
+}
+
+func getElasticIPsLimit(sess client.ConfigProvider) int {
+	ec2Client := ec2.New(sess)
+	input := &ec2.DescribeAccountAttributesInput{
+		AttributeNames: []*string{
+			aws.String("vpc-max-elastic-ips"),
+		},
+	}
+	result, err := ec2Client.DescribeAccountAttributes(input)
+	Expect(err).NotTo(HaveOccurred())
+	res, err := strconv.Atoi(*result.AccountAttributes[0].AttributeValues[0].AttributeValue)
+	Expect(err).NotTo(HaveOccurred())
+	return res
+}
+
+func getEvents(namespace string) *corev1.EventList {
+	eventsList := &corev1.EventList{}
+	if err := bootstrapClusterProxy.GetClient().List(context.TODO(), eventsList, crclient.InNamespace(namespace), crclient.MatchingLabels{}); err != nil {
+		fmt.Fprintf(GinkgoWriter, "Got error while fetching events of namespace: %s, %s \n", namespace, err.Error())
+	}
+
+	return eventsList
+}
+
+func getMachinesOfDeployment(namespace *corev1.Namespace, machineDeployment *clusterv1.MachineDeployment) *clusterv1.MachineList {
+	Byf("Fetching the Machines of MachineDeployment %s", machineDeployment.Name)
+	machineList := &clusterv1.MachineList{}
+	k8sClient := bootstrapClusterProxy.GetClient()
+
+	selector, err := metav1.LabelSelectorAsMap(&machineDeployment.Spec.Selector)
 	Expect(err).NotTo(HaveOccurred())
 
-	kubeConfigPath := path.Join(tmpDir, clusterName+".kubeconfig")
-	Expect(ioutil.WriteFile(kubeConfigPath, kubeConfigData, 0o640)).To(Succeed())
-
-	kubeConfigData, readErr := ioutil.ReadFile(kubeConfigPath)
-	Expect(readErr).NotTo(HaveOccurred())
-
-	restConfig, err := clientcmd.RESTConfigFromKubeConfig(kubeConfigData)
+	err = k8sClient.List(context.TODO(), machineList, crclient.InNamespace(namespace.Name), crclient.MatchingLabels(selector))
 	Expect(err).NotTo(HaveOccurred())
+	return machineList
+}
 
-	k8sclient, err := crclient.New(restConfig, crclient.Options{})
+func getSession() client.ConfigProvider {
+	sess, err := session.NewSessionWithOptions(session.Options{
+		SharedConfigState: session.SharedConfigEnable,
+	})
 	Expect(err).NotTo(HaveOccurred())
-
-	return kubeConfigPath, k8sclient
+	return sess
 }
 
 func getSubnetId(filterKey, filterValue string) *string {
-	ec2c := ec2.New(sess)
+	var subnetOutput *ec2.DescribeSubnetsOutput
+	var err error
+
+	ec2Client := ec2.New(getSession())
 	subnetInput := &ec2.DescribeSubnetsInput{
 		Filters: []*ec2.Filter{
 			{
@@ -953,115 +681,41 @@ func getSubnetId(filterKey, filterValue string) *string {
 			},
 		},
 	}
-	result, err := ec2c.DescribeSubnets(subnetInput)
+
+	Eventually(func() int {
+		subnetOutput, err = ec2Client.DescribeSubnets(subnetInput)
+		Expect(err).NotTo(HaveOccurred())
+		return len(subnetOutput.Subnets)
+	}, e2eConfig.GetIntervals("", "wait-infra-subnets")...).Should(Equal(1))
+
+	return subnetOutput.Subnets[0].SubnetId
+}
+
+func getVolumeIds(info statefulSetInfo, k8sclient crclient.Client) []*string {
+	By("Retrieving IDs of dynamically provisioned volumes.")
+	statefulset := &appsv1.StatefulSet{}
+	err := k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: info.name}, statefulset)
 	Expect(err).NotTo(HaveOccurred())
-	return result.Subnets[0].SubnetId
-}
-
-func getVpcId(vpcName string) string {
-	ec2c := ec2.New(sess)
-	input := &ec2.DescribeVpcsInput{}
-	result, err := ec2c.DescribeVpcs(input)
+	podSelector, err := metav1.LabelSelectorAsMap(statefulset.Spec.Selector)
+	pvcList := &corev1.PersistentVolumeClaimList{}
+	err = k8sclient.List(context.TODO(), pvcList, crclient.InNamespace(info.namespace), crclient.MatchingLabels(podSelector))
 	Expect(err).NotTo(HaveOccurred())
-	for _, tmpVpc := range result.Vpcs {
-		for _, tmpTag := range tmpVpc.Tags {
-			if *tmpTag.Key == "Name" && *tmpTag.Value == vpcName {
-				return *tmpVpc.VpcId
-			}
-		}
+	var volIds []*string
+	for _, pvc := range pvcList.Items {
+		volName := pvc.Spec.VolumeName
+		volDescription := &corev1.PersistentVolume{}
+		err = k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: volName}, volDescription)
+		Expect(err).NotTo(HaveOccurred())
+		urlSlice := strings.Split(volDescription.Spec.AWSElasticBlockStore.VolumeID, "/")
+		volIds = append(volIds, &urlSlice[len(urlSlice)-1])
 	}
-	return ""
-}
-
-func verfiyInstancesInSubnet(numOfInstances int32, subnetId *string) {
-	ec2c := ec2.New(sess)
-	instanceInput := &ec2.DescribeInstancesInput{
-		Filters: []*ec2.Filter{
-			{
-				Name:   aws.String("subnet-id"),
-				Values: []*string{subnetId},
-			},
-			{
-				Name:   aws.String("instance-state-name"),
-				Values: []*string{aws.String("running")},
-			},
-		},
-	}
-	result, err := ec2c.DescribeInstances(instanceInput)
-	Expect(err).NotTo(HaveOccurred())
-	Expect(int(numOfInstances)).To(Equal(len(result.Reservations)))
-}
-
-func getAvailabilityZone() []*ec2.AvailabilityZone {
-	if sess == nil {
-		panic("session is nil, cant get availability zones.")
-	}
-	ec2c := ec2.New(sess)
-	azs, err := ec2c.DescribeAvailabilityZones(nil)
-	Expect(err).NotTo(HaveOccurred())
-	return azs.AvailabilityZones
-}
-
-func createCluster(namespace, clusterName, awsClusterName string, multiAZ bool) bool {
-	By("Creating an AWSCluster")
-	makeAWSCluster(namespace, awsClusterName, multiAZ)
-
-	By("Creating a Cluster")
-	makeCluster(namespace, clusterName, awsClusterName)
-
-	By("Ensuring Cluster Infrastructure Reports as Ready")
-	return waitForClusterInfrastructureReady(namespace, clusterName)
-}
-
-func makeSingleControlPlaneCluster(setup testSetup) crclient.Client {
-	if setup.namespace == "" && setup.clusterName == "" && setup.awsClusterName == "" {
-		panic(" setup looks effectively nil /  uninitialized... ")
-	}
-	Expect(createCluster(setup.namespace, setup.clusterName, setup.awsClusterName, setup.multipleAZ)).Should(BeTrue())
-
-	By("Creating the initial Control Plane Machine")
-	awsMachineName := setup.cpAWSMachinePrefix + "-0"
-	bootstrapConfigName := setup.cpBootstrapConfigPrefix + "-0"
-	machineName := setup.cpMachinePrefix + "-0"
-	createInitialControlPlaneMachine(setup, machineName, awsMachineName, bootstrapConfigName)
-	clusterKubeConfigPath, clusterK8sClient := createClusterKubeConfigs(setup.testTmpDir, setup.namespace, setup.clusterName)
-
-	By("Deploying CNI to created Cluster")
-	deployCNI(clusterKubeConfigPath, *cniManifests)
-	waitForMachineNodeReady(setup.namespace, machineName)
-	return clusterK8sClient
-}
-
-func createPendingMachineDeployment(setup testSetup) {
-	fmt.Fprintf(GinkgoWriter, "Creating MachineDeployment in namespace %s with name %s\n", setup.namespace, setup.machineDeploymentName)
-	makeAWSMachineTemplate(setup.namespace, setup.awsMachineTemplateName, setup.instanceType, setup.availabilityZone, setup.subnetId)
-	makeJoinBootstrapConfigTemplate(setup.namespace, setup.mdBootstrapConfig)
-	makeMachineDeployment(setup.namespace, setup.machineDeploymentName, setup.awsMachineTemplateName, setup.mdBootstrapConfig, setup.clusterName, setup.initialReplicas)
-}
-
-func scaleMachineDeployment(namespace, machineDeployment string, replicasCurrent int32, replicasProposed int32) {
-	fmt.Fprintf(GinkgoWriter, "Scaling MachineDeployment from %d to %d\n", replicasCurrent, replicasProposed)
-	deployment := &clusterv1.MachineDeployment{}
-	Expect(kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: machineDeployment}, deployment))
-	deployment.Spec.Replicas = &replicasProposed
-	Expect(kindClient.Update(context.TODO(), deployment)).NotTo(HaveOccurred())
-	waitForMachinesCountMatch(namespace, machineDeployment, replicasCurrent, replicasProposed)
-	waitForMachineDeploymentRunning(namespace, machineDeployment)
-}
-
-func createMachineDeployment(setup testSetup) {
-	fmt.Fprintf(GinkgoWriter, "Creating MachineDeployment in namespace %s with name %s\n", setup.namespace, setup.machineDeploymentName)
-	makeAWSMachineTemplate(setup.namespace, setup.awsMachineTemplateName, setup.instanceType, setup.availabilityZone, setup.subnetId)
-	makeJoinBootstrapConfigTemplate(setup.namespace, setup.mdBootstrapConfig)
-	makeMachineDeployment(setup.namespace, setup.machineDeploymentName, setup.awsMachineTemplateName, setup.mdBootstrapConfig, setup.clusterName, setup.initialReplicas)
-	waitForMachinesCountMatch(setup.namespace, setup.machineDeploymentName, setup.initialReplicas, setup.initialReplicas)
-	waitForMachineDeploymentRunning(setup.namespace, setup.machineDeploymentName)
+	return volIds
 }
 
 func isErrorEventExists(namespace, machineDeploymentName, eventReason, errorMsg string, eList *corev1.EventList) bool {
-	fmt.Fprintf(GinkgoWriter, "Checking Error event with message %s is present \n", errorMsg)
+	k8sClient := bootstrapClusterProxy.GetClient()
 	machineDeployment := &clusterv1.MachineDeployment{}
-	if err := kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineDeploymentName}, machineDeployment); err != nil {
+	if err := k8sClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineDeploymentName}, machineDeployment); err != nil {
 		fmt.Fprintf(GinkgoWriter, "Got error while getting machinedeployment %s \n", machineDeploymentName)
 		return false
 	}
@@ -1073,7 +727,7 @@ func isErrorEventExists(namespace, machineDeploymentName, eventReason, errorMsg 
 	}
 
 	awsMachineList := &infrav1.AWSMachineList{}
-	if err := kindClient.List(context.TODO(), awsMachineList, crclient.InNamespace(namespace), crclient.MatchingLabels(selector)); err != nil {
+	if err := k8sClient.List(context.TODO(), awsMachineList, crclient.InNamespace(namespace), crclient.MatchingLabels(selector)); err != nil {
 		fmt.Fprintf(GinkgoWriter, "Got error while getting awsmachines of machinedeployment: %s, %s \n", machineDeploymentName, err.Error())
 		return false
 	}
@@ -1093,93 +747,59 @@ func isErrorEventExists(namespace, machineDeploymentName, eventReason, errorMsg 
 	return false
 }
 
-func getEvents(namespace string) *corev1.EventList {
-	eventsList := &corev1.EventList{}
-	if err := kindClient.List(context.TODO(), eventsList, crclient.InNamespace(namespace), crclient.MatchingLabels{}); err != nil {
-		fmt.Fprintf(GinkgoWriter, "Got error while fetching events of namespace: %s, %s \n", namespace, err.Error())
-	}
-
-	return eventsList
-}
-
-func deleteMachine(namespace, machineDeploymentName string) {
-	machineDeployment := &clusterv1.MachineDeployment{}
-	Expect(kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineDeploymentName}, machineDeployment)).To(Succeed())
-	machineList := &clusterv1.MachineList{}
-
-	selector, err := metav1.LabelSelectorAsMap(&machineDeployment.Spec.Selector)
-	Expect(err).NotTo(HaveOccurred())
-
-	err = kindClient.List(context.TODO(), machineList, crclient.InNamespace(namespace), crclient.MatchingLabels(selector))
-	Expect(err).NotTo(HaveOccurred())
-
-	Expect(len(machineList.Items)).ToNot(Equal(0))
-	machine := &clusterv1.Machine{
+func makeAWSMachineTemplate(namespace, name, instanceType string, az, subnetId *string) *infrav1.AWSMachineTemplate {
+	awsMachine := &infrav1.AWSMachineTemplate{
 		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
 			Namespace: namespace,
-			Name:      machineList.Items[0].Name,
+		},
+		Spec: infrav1.AWSMachineTemplateSpec{
+			Template: infrav1.AWSMachineTemplateResource{
+				Spec: infrav1.AWSMachineSpec{
+					InstanceType:       instanceType,
+					IAMInstanceProfile: "nodes.cluster-api-provider-aws.sigs.k8s.io",
+					SSHKeyName:         pointer.StringPtr(os.Getenv("AWS_SSH_KEY_NAME")),
+				},
+			},
 		},
 	}
-	Expect(kindClient.Delete(context.TODO(), machine)).To(Succeed())
-}
-
-func waitForMachinesCountMatch(namespace, machineDeploymentName string, replicasCurrent int32, replicasProposed int32) {
-	fmt.Fprintf(GinkgoWriter, "Ensuring Machine count matched to %d \n", replicasProposed)
-	machineDeployment := &clusterv1.MachineDeployment{}
-	if err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: machineDeploymentName}, machineDeployment); err != nil {
-		return
+	if az != nil {
+		awsMachine.Spec.Template.Spec.FailureDomain = az
 	}
-	scalingDown := replicasCurrent > replicasProposed
-	Eventually(
-		func() (int32, error) {
-			machineList := &clusterv1.MachineList{}
-			selector, err := metav1.LabelSelectorAsMap(&machineDeployment.Spec.Selector)
-			if err != nil {
-				return -1, err
-			}
-			if err := kindClient.List(context.TODO(), machineList, crclient.InNamespace(namespace), crclient.MatchingLabels(selector)); err != nil {
-				return -1, err
-			} else {
-				var runningMachines int32
-				var deletingMachines int32
-				for _, item := range machineList.Items {
-					if string(clusterv1.MachinePhaseRunning) == item.Status.Phase {
-						runningMachines += 1
-					}
-					if string(clusterv1.MachinePhaseDeleting) == item.Status.Phase {
-						deletingMachines += 1
-					}
-				}
-				if scalingDown {
-					if deletingMachines > 0 {
-						return -1, nil
-					}
-				}
-				return runningMachines, nil
-			}
-		},
-		20*time.Minute, 30*time.Second,
-	).Should(Equal(replicasProposed))
-}
 
-func waitForMachineDeploymentRunning(namespace, machineDeploymentName string) bool {
-	fmt.Fprintf(GinkgoWriter, "Verifying MachineDeployment %s/%s is running..\n", namespace, machineDeploymentName)
-	endTime := time.Now().Add(5 * time.Minute)
-	for time.Now().Before(endTime) {
-		machineDeployment := &clusterv1.MachineDeployment{}
-		if err := kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: machineDeploymentName}, machineDeployment); nil == err {
-			if *machineDeployment.Spec.Replicas == machineDeployment.Status.ReadyReplicas {
-				return true
-			}
+	if subnetId != nil {
+		resRef := &infrav1.AWSResourceReference{
+			ID: subnetId,
 		}
-		time.Sleep(15 * time.Second)
+		awsMachine.Spec.Template.Spec.Subnet = resRef
 	}
-	return false
+
+	return awsMachine
 }
 
-func makeMachineDeployment(namespace, mdName, awsMachineTemplateName, bootstrapConfigName, clusterName string, replicas int32) {
-	fmt.Fprintf(GinkgoWriter, "Creating MachineDeployment %s/%s\n", namespace, mdName)
-	machineDeployment := &clusterv1.MachineDeployment{
+func makeJoinBootstrapConfigTemplate(namespace, name string) *bootstrapv1.KubeadmConfigTemplate {
+	return &bootstrapv1.KubeadmConfigTemplate{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: bootstrapv1.KubeadmConfigTemplateSpec{
+			Template: bootstrapv1.KubeadmConfigTemplateResource{
+				Spec: bootstrapv1.KubeadmConfigSpec{
+					JoinConfiguration: &kubeadmv1beta1.JoinConfiguration{
+						NodeRegistration: kubeadmv1beta1.NodeRegistrationOptions{
+							Name:             "{{ ds.meta_data.local_hostname }}",
+							KubeletExtraArgs: map[string]string{"cloud-provider": "aws"},
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func makeMachineDeployment(namespace, mdName, clusterName string, replicas int32) *clusterv1.MachineDeployment {
+	return &clusterv1.MachineDeployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      mdName,
 			Namespace: namespace,
@@ -1210,518 +830,79 @@ func makeMachineDeployment(namespace, mdName, awsMachineTemplateName, bootstrapC
 						ConfigRef: &corev1.ObjectReference{
 							Kind:       "KubeadmConfigTemplate",
 							APIVersion: bootstrapv1.GroupVersion.String(),
-							Name:       bootstrapConfigName,
+							Name:       mdName,
 							Namespace:  namespace,
 						},
 					},
 					InfrastructureRef: corev1.ObjectReference{
 						Kind:       "AWSMachineTemplate",
 						APIVersion: infrav1.GroupVersion.String(),
-						Name:       awsMachineTemplateName,
+						Name:       mdName,
 						Namespace:  namespace,
 					},
-					Version: k8sVersion,
+					Version: pointer.StringPtr(e2eConfig.GetVariable(KubernetesVersion)),
 				},
 			},
 		},
 	}
-	Expect(kindClient.Create(context.TODO(), machineDeployment)).NotTo(HaveOccurred())
 }
 
-func makeJoinBootstrapConfigTemplate(namespace, name string) {
-	fmt.Fprintf(GinkgoWriter, "Creating Join KubeadmConfigTemplate %s/%s\n", namespace, name)
-	configTemplate := &bootstrapv1.KubeadmConfigTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: bootstrapv1.KubeadmConfigTemplateSpec{
-			Template: bootstrapv1.KubeadmConfigTemplateResource{
-				Spec: bootstrapv1.KubeadmConfigSpec{
-					JoinConfiguration: &kubeadmv1beta1.JoinConfiguration{
-						NodeRegistration: kubeadmv1beta1.NodeRegistrationOptions{
-							Name:             "{{ ds.meta_data.local_hostname }}",
-							KubeletExtraArgs: map[string]string{"cloud-provider": "aws"},
-						},
-					},
-				},
-			},
+func terminateInstance(instanceId string) {
+	Byf("Terminating EC2 instance with ID: %s", instanceId)
+	ec2Client := ec2.New(getSession())
+	input := &ec2.TerminateInstancesInput{
+		InstanceIds: []*string{
+			aws.String(instanceId[strings.LastIndex(instanceId, "/")+1:]),
 		},
 	}
-	Expect(kindClient.Create(context.TODO(), configTemplate)).NotTo(HaveOccurred())
+
+	result, err := ec2Client.TerminateInstances(input)
+	Expect(err).To(BeNil())
+	Expect(len(result.TerminatingInstances)).To(Equal(1))
+	termCode := int64(32)
+	Expect(*result.TerminatingInstances[0].CurrentState.Code).To(Equal(termCode))
 }
 
-func makeAWSMachineTemplate(namespace, name, instanceType string, az, subnetId *string) {
-	fmt.Fprintf(GinkgoWriter, "Creating AWSMachineTemplate %s/%s\n", namespace, name)
-	awsMachine := &infrav1.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: infrav1.AWSMachineTemplateSpec{
-			Template: infrav1.AWSMachineTemplateResource{
-				Spec: infrav1.AWSMachineSpec{
-					InstanceType:       instanceType,
-					IAMInstanceProfile: "nodes.cluster-api-provider-aws.sigs.k8s.io",
-					SSHKeyName:         pointer.StringPtr(keyPairName),
-				},
-			},
+func verifyElbExists(elbName string, exists bool) {
+	Byf("Verifying ELB with name %s present", elbName)
+	elbClient := elb.New(getSession())
+	input := &elb.DescribeLoadBalancersInput{
+		LoadBalancerNames: []*string{
+			aws.String(elbName),
 		},
 	}
-	if az != nil {
-		awsMachine.Spec.Template.Spec.FailureDomain = az
+	elbsOutput, err := elbClient.DescribeLoadBalancers(input)
+	if exists {
+		Expect(err).NotTo(HaveOccurred())
+		Expect(len(elbsOutput.LoadBalancerDescriptions)).To(Equal(1))
+		Byf("ELB with name %s exists", elbName)
+	} else {
+		aerr, ok := err.(awserr.Error)
+		Expect(ok).To(BeTrue())
+		Expect(aerr.Code()).To(Equal(elb.ErrCodeAccessPointNotFoundException))
+		Byf("ELB with name %s doesn't exists", elbName)
 	}
+}
 
-	if subnetId != nil {
-		resRef := &infrav1.AWSResourceReference{
-			ID: subnetId,
-		}
-		awsMachine.Spec.Template.Spec.Subnet = resRef
+func verifyVolumesExists(awsVolumeIds []*string) {
+	By("Ensuring dynamically provisioned volumes exists")
+	ec2Client := ec2.New(getSession())
+	input := &ec2.DescribeVolumesInput{
+		VolumeIds: awsVolumeIds,
 	}
-	Expect(kindClient.Create(context.TODO(), awsMachine)).NotTo(HaveOccurred())
-}
-
-func deployCNI(kubeConfigPath, manifestPath string) {
-	Expect(exec.Command(
-		*kubectlBinary,
-		"apply",
-		"--kubeconfig",
-		kubeConfigPath,
-		"-f",
-		manifestPath,
-	).Run()).To(Succeed())
-}
-
-func waitForWorkerAPIServerReady(namespace, clusterName string) {
-	kubeConfigData, err := kubeconfig.FromSecret(context.TODO(), kindClient, crclient.ObjectKey{Name: clusterName, Namespace: namespace})
+	_, err := ec2Client.DescribeVolumes(input)
 	Expect(err).NotTo(HaveOccurred())
-
-	restConfig, err := clientcmd.RESTConfigFromKubeConfig(kubeConfigData)
-	Expect(err).NotTo(HaveOccurred())
-
-	nodeClient, err := crclient.New(restConfig, crclient.Options{})
-	Expect(err).NotTo(HaveOccurred())
-
-	Eventually(
-		func() bool {
-			nodes := &corev1.NodeList{}
-			if err := nodeClient.List(context.TODO(), nodes); err != nil {
-				fmt.Fprintf(GinkgoWriter, "(waitForWorkerAPIServerReady) Error retrieving nodes: %v", err)
-				return false
-			}
-			return true
-		},
-		5*time.Minute, 15*time.Second,
-	).Should(BeTrue())
 }
 
-func waitForMachineNodeReady(namespace, name string) {
-	machine := &clusterv1.Machine{}
-	Expect(kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: name}, machine)).To(Succeed())
-
-	nodeName := machine.Status.NodeRef.Name
-
-	cluster := crclient.ObjectKey{
-		Namespace: namespace,
-		Name:      machine.Labels[clusterv1.ClusterLabelName],
-	}
-
-	kubeConfigData, err := kubeconfig.FromSecret(context.TODO(), kindClient, cluster)
-	Expect(err).NotTo(HaveOccurred())
-
-	restConfig, err := clientcmd.RESTConfigFromKubeConfig(kubeConfigData)
-	Expect(err).NotTo(HaveOccurred())
-
-	nodeClient, err := crclient.New(restConfig, crclient.Options{})
-	Expect(err).NotTo(HaveOccurred())
-
-	Eventually(
-		func() bool {
-			node := &corev1.Node{}
-			if err := nodeClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Name: nodeName}, node); err != nil {
-				fmt.Fprintf(GinkgoWriter, "(waitForMachineNodeReady) Error retrieving node: %v \n", err)
-				return false
-			}
-			conditionMap := make(map[corev1.NodeConditionType]*corev1.NodeCondition)
-			for i, condition := range node.Status.Conditions {
-				if condition.Type == corev1.NodeReady {
-					conditionMap[condition.Type] = &node.Status.Conditions[i]
-				}
-			}
-
-			if condition, ok := conditionMap[corev1.NodeReady]; ok {
-				if condition.Status == corev1.ConditionTrue {
-					fmt.Fprintf(GinkgoWriter, "(waitForMachineNodeReady) Success retrieving node: NodeReady %v \n", err)
-					return true
-				}
-			}
-			fmt.Fprintf(GinkgoWriter, "(waitForMachineNodeReady) machine still not ready: %v \n", err)
-
-			return false
-		},
-		2*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func createAdditionalControlPlaneMachine(setup testSetup, machineName, awsMachineName, bootstrapConfigName string) {
-	makeAWSMachine(setup.namespace, awsMachineName, setup.instanceType, setup.availabilityZone, setup.subnetId)
-	makeJoinBootstrapConfig(setup.namespace, bootstrapConfigName)
-	makeMachine(setup.namespace, machineName, awsMachineName, bootstrapConfigName, setup.clusterName)
-	waitForMachineBootstrapReady(setup.namespace, machineName)
-	waitForAWSMachineRunning(setup.namespace, awsMachineName)
-	waitForAWSMachineReady(setup.namespace, awsMachineName)
-	waitForMachineNodeRef(setup.namespace, machineName)
-	waitForMachineNodeReady(setup.namespace, machineName)
-}
-
-func createInitialControlPlaneMachine(setup testSetup, machineName, awsMachineName, bootstrapConfigName string) {
-	makeAWSMachine(setup.namespace, awsMachineName, setup.instanceType, setup.availabilityZone, setup.subnetId)
-	makeInitBootstrapConfig(setup.namespace, bootstrapConfigName)
-	makeMachine(setup.namespace, machineName, awsMachineName, bootstrapConfigName, setup.clusterName)
-	waitForMachineBootstrapReady(setup.namespace, machineName)
-	waitForAWSMachineRunning(setup.namespace, awsMachineName)
-	waitForAWSMachineReady(setup.namespace, awsMachineName)
-	waitForMachineNodeRef(setup.namespace, machineName)
-	waitForClusterControlPlaneInitialized(setup.namespace, setup.clusterName)
-}
-
-func deleteCluster(namespace, name string) {
-	fmt.Fprintf(GinkgoWriter, "Deleting Cluster %s/%s\n", namespace, name)
-	Expect(kindClient.Delete(context.TODO(), &clusterv1.Cluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: namespace,
-			Name:      name,
-		},
-	})).To(Succeed())
-
-	Eventually(
-		func() bool {
-			cluster := &clusterv1.Cluster{}
-			err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, cluster)
-			fmt.Fprintf(GinkgoWriter, "Eventually (deleteCluster) --> %s/%s : Status = %v \n", namespace, name, err)
-			// non nil error ==> cluster is gone, which means delete worked.
-			if err != nil {
-				if apierrors.IsNotFound(err) {
-					return true
-				}
-				return false
-			}
-			return false
-		},
-		20*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func waitForMachineNodeRef(namespace, name string) {
-	Eventually(
-		func() *corev1.ObjectReference {
-			machine := &clusterv1.Machine{}
-			defer func() {
-				fmt.Fprintf(GinkgoWriter, "Eventually (waitForMachineNodeRef) --> %s/%s : Status = %v \n", namespace, name, machine.Status)
-			}()
-			if err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, machine); err != nil {
-				return nil
-			}
-			return machine.Status.NodeRef
-		},
-		10*time.Minute, 15*time.Second,
-	).ShouldNot(BeNil())
-}
-
-func waitForClusterControlPlaneInitialized(namespace, name string) {
-	fmt.Fprintf(GinkgoWriter, "Ensuring control plane initialized for cluster %s/%s\n", namespace, name)
+func waitForStatefulSetRunning(info statefulSetInfo, k8sclient crclient.Client) {
+	Byf("Ensuring Statefulset(%s) is running", info.name)
+	statefulset := &appsv1.StatefulSet{}
 	Eventually(
 		func() (bool, error) {
-			cluster := &clusterv1.Cluster{}
-			defer func() {
-				fmt.Fprintf(GinkgoWriter, "Eventually (waitForClusterControlPlaneInitialized) --> %s/%s : Status = %v \n", namespace, name, cluster)
-			}()
-
-			if err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, cluster); err != nil {
+			if err := k8sclient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: info.namespace, Name: info.name}, statefulset); err != nil {
 				return false, err
 			}
-			return cluster.Status.ControlPlaneInitialized, nil
-		},
-		15*time.Minute, 15*time.Second,
+			return *statefulset.Spec.Replicas == statefulset.Status.ReadyReplicas, nil
+		}, 10*time.Minute, 30*time.Second,
 	).Should(BeTrue())
-}
-
-func waitForAWSMachineRunning(namespace, name string) {
-	Eventually(
-		func() (bool, error) {
-			var err error
-			awsMachine := &infrav1.AWSMachine{}
-			defer func() {
-				fmt.Fprintf(GinkgoWriter, "Eventually (waitForAWSMachineRunning) --> %s/%s : Status = %v \n", namespace, name, awsMachine)
-			}()
-			if err = kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, awsMachine); err != nil {
-				return false, err
-			}
-			if awsMachine.Status.InstanceState == nil {
-				return false, nil
-			}
-			return *awsMachine.Status.InstanceState == infrav1.InstanceStateRunning, nil
-		},
-		5*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func waitForClusterInfrastructureReady(namespace, name string) bool {
-	fmt.Fprintf(GinkgoWriter, "Ensuring infrastructure is ready for cluster %s/%s\n", namespace, name)
-	polls := 0
-	var err error
-	endTime := time.Now().Add(20 * time.Minute)
-	for time.Now().Before(endTime) {
-		polls = polls + 1
-		cluster := &clusterv1.Cluster{}
-
-		logit := func() {
-			fmt.Fprintf(GinkgoWriter, "POLLING ( # %v ) (waitForClusterInfrastructureReady) --> Ensuring infrastructure is ready for cluster %s/%s : Status = %+v , Err = %v \n", polls, namespace, name, cluster.Status, err)
-		}
-
-		if err = kindClient.Get(context.TODO(), apimachinerytypes.NamespacedName{Namespace: namespace, Name: name}, cluster); nil == err {
-			if cluster.Status.InfrastructureReady {
-				logit()
-				return true
-			}
-		}
-		if polls == 1 || polls%5 == 0 {
-			logit()
-		}
-		time.Sleep(15 * time.Second)
-	}
-	fmt.Fprintf(GinkgoWriter, "Ensuring infrastructure is ready for cluster <<<< GIVING UP, TIME EXPIRED >>>> %v %v %v n", namespace, name, err)
-
-	return false
-}
-
-func waitForMachineBootstrapReady(namespace, name string) {
-	Eventually(
-		func() (bool, error) {
-			machine := &clusterv1.Machine{}
-			err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, machine)
-			fmt.Fprintf(GinkgoWriter, "Eventually (waitForMachineBootstrapReady) -> Ensuring Machine %s/%s has bootstrapReady... error = %v \n", namespace, name, err)
-			if err != nil {
-				return false, err
-			} else {
-				return machine.Status.BootstrapReady, nil
-			}
-		},
-		5*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func waitForAWSMachineReady(namespace, name string) {
-	Eventually(
-		func() (bool, error) {
-			awsMachine := &infrav1.AWSMachine{}
-			err := kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: name}, awsMachine)
-			fmt.Fprintf(GinkgoWriter, "Eventually (waitForAWSMachineReady) -> Ensuring Machine %s/%s has bootstrapReady... error = %v \n", namespace, name, err)
-			if err != nil {
-				return false, err
-			} else {
-				return awsMachine.Status.Ready, nil
-			}
-		},
-		2*time.Minute, 15*time.Second,
-	).Should(BeTrue())
-}
-
-func makeMachine(namespace, name, awsMachineName, bootstrapConfigName, clusterName string) {
-	fmt.Fprintf(GinkgoWriter, "Creating Machine %s/%s\n", namespace, name)
-	machine := &clusterv1.Machine{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-			Labels: map[string]string{
-				clusterv1.MachineControlPlaneLabelName: "true",
-			},
-		},
-		Spec: clusterv1.MachineSpec{
-			ClusterName: clusterName,
-			Bootstrap: clusterv1.Bootstrap{
-				ConfigRef: &corev1.ObjectReference{
-					Kind:       "KubeadmConfig",
-					APIVersion: bootstrapv1.GroupVersion.String(),
-					Name:       bootstrapConfigName,
-					Namespace:  namespace,
-				},
-			},
-			InfrastructureRef: corev1.ObjectReference{
-				Kind:       "AWSMachine",
-				APIVersion: infrav1.GroupVersion.String(),
-				Name:       awsMachineName,
-				Namespace:  namespace,
-			},
-			Version: k8sVersion,
-		},
-	}
-	Expect(kindClient.Create(context.TODO(), machine)).To(Succeed())
-}
-
-func makeJoinBootstrapConfig(namespace, name string) {
-	fmt.Fprintf(GinkgoWriter, "Creating Join KubeadmConfig %s/%s\n", namespace, name)
-	config := &bootstrapv1.KubeadmConfig{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: bootstrapv1.KubeadmConfigSpec{
-			JoinConfiguration: &kubeadmv1beta1.JoinConfiguration{
-				ControlPlane: &kubeadmv1beta1.JoinControlPlane{},
-				NodeRegistration: kubeadmv1beta1.NodeRegistrationOptions{
-					Name:             "{{ ds.meta_data.local_hostname }}",
-					KubeletExtraArgs: map[string]string{"cloud-provider": "aws"},
-				},
-			},
-		},
-	}
-	Expect(kindClient.Create(context.TODO(), config)).To(Succeed())
-}
-
-func makeInitBootstrapConfig(namespace, name string) {
-	fmt.Fprintf(GinkgoWriter, "Creating Init KubeadmConfig %s/%s\n", namespace, name)
-	config := &bootstrapv1.KubeadmConfig{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: bootstrapv1.KubeadmConfigSpec{
-			ClusterConfiguration: &kubeadmv1beta1.ClusterConfiguration{
-				APIServer: kubeadmv1beta1.APIServer{
-					ControlPlaneComponent: kubeadmv1beta1.ControlPlaneComponent{
-						ExtraArgs: map[string]string{"cloud-provider": "aws"},
-					},
-				},
-				ControllerManager: kubeadmv1beta1.ControlPlaneComponent{
-					ExtraArgs: map[string]string{"cloud-provider": "aws"},
-				},
-			},
-			InitConfiguration: &kubeadmv1beta1.InitConfiguration{
-				NodeRegistration: kubeadmv1beta1.NodeRegistrationOptions{
-					Name:             "{{ ds.meta_data.local_hostname }}",
-					KubeletExtraArgs: map[string]string{"cloud-provider": "aws"},
-				},
-			},
-			Files: []bootstrapv1.File{
-				{
-					Path:    "/tmp/userdata-length-test.txt",
-					Content: apirand.String(20000),
-				},
-			},
-		},
-	}
-	Expect(kindClient.Create(context.TODO(), config)).To(Succeed())
-}
-
-func makeAWSMachine(namespace, name, instanceType string, az, subnetId *string) {
-	fmt.Fprintf(GinkgoWriter, "Creating AWSMachine %s/%s\n", namespace, name)
-	awsMachine := &infrav1.AWSMachine{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: infrav1.AWSMachineSpec{
-			InstanceType:       instanceType,
-			IAMInstanceProfile: "control-plane.cluster-api-provider-aws.sigs.k8s.io",
-			SSHKeyName:         pointer.StringPtr(keyPairName),
-		},
-	}
-	if az != nil {
-		awsMachine.Spec.FailureDomain = az
-	}
-	if subnetId != nil {
-		awsMachine.Spec.Subnet = &infrav1.AWSResourceReference{ID: subnetId}
-	}
-	Expect(kindClient.Create(context.TODO(), awsMachine)).To(Succeed())
-}
-
-func makeCluster(namespace, name, awsClusterName string) {
-	fmt.Fprintf(GinkgoWriter, "Creating Cluster %s/%s\n", namespace, name)
-	cluster := &clusterv1.Cluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: clusterv1.ClusterSpec{
-			ClusterNetwork: &clusterv1.ClusterNetwork{
-				Pods: &clusterv1.NetworkRanges{
-					CIDRBlocks: []string{"192.168.0.0/16"},
-				},
-			},
-			InfrastructureRef: &corev1.ObjectReference{
-				Kind:       "AWSCluster",
-				APIVersion: infrav1.GroupVersion.String(),
-				Name:       awsClusterName,
-				Namespace:  namespace,
-			},
-		},
-	}
-	Expect(kindClient.Create(context.TODO(), cluster)).To(Succeed())
-}
-
-func makeAWSCluster(namespace, name string, multipleAZ bool) {
-	fmt.Fprintf(GinkgoWriter, "Creating AWSCluster %s/%s multi %v \n", namespace, name, multipleAZ)
-	awsCluster := &infrav1.AWSCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-		Spec: infrav1.AWSClusterSpec{
-			Region:     region,
-			SSHKeyName: pointer.StringPtr(keyPairName),
-			NetworkSpec: infrav1.NetworkSpec{
-				VPC: infrav1.VPCSpec{
-					AvailabilityZoneUsageLimit: gophercloud.IntToPointer(1),
-				},
-			},
-		},
-	}
-	if multipleAZ {
-		// This call can panic ...
-		azs := getAvailabilityZone()
-		availabilityZones = append(availabilityZones, azs[0].ZoneName, azs[1].ZoneName, azs[2].ZoneName)
-		subnets := []*infrav1.SubnetSpec{
-			{
-				CidrBlock:        privateCIDRs[0],
-				AvailabilityZone: *availabilityZones[0],
-			},
-			{
-				CidrBlock:        publicCIDRs[0],
-				AvailabilityZone: *availabilityZones[0],
-				IsPublic:         true,
-			},
-			{
-				CidrBlock:        privateCIDRs[1],
-				AvailabilityZone: *availabilityZones[1],
-			},
-			{
-				CidrBlock:        publicCIDRs[1],
-				AvailabilityZone: *availabilityZones[1],
-				IsPublic:         true,
-			},
-			{
-				CidrBlock:        privateCIDRs[2],
-				AvailabilityZone: *availabilityZones[2],
-			},
-			{
-				CidrBlock:        publicCIDRs[2],
-				AvailabilityZone: *availabilityZones[2],
-				IsPublic:         true,
-			},
-		}
-		awsCluster.Spec.NetworkSpec.Subnets = subnets
-	}
-
-	Expect(kindClient.Create(context.TODO(), awsCluster)).To(Succeed())
-}
-
-func createNamespace(namespace string) {
-	fmt.Fprintf(GinkgoWriter, "creating namespace %q\n", namespace)
-	ns := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: namespace,
-		},
-	}
-	Expect(kindClient.Create(context.TODO(), ns)).To(Succeed())
 }
