+++ test/e2e/data/cni/calico.yaml
