+++ test/e2e/data/infrastructure-aws/cluster-template-limit-az.yaml
