+++ test/e2e/data/infrastructure-aws/cluster-template-multi-az.yaml
