+++ test/e2e/data/kubetest/kustomization/kustomization.yaml
