+++ test/e2e/data/kubetest/kustomization/kustomizeversions.yaml
