+++ test/e2e/e2e_suite_test.go
@@ -1,7 +1,7 @@
 // +build e2e
 
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2020 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,311 +16,403 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_test
+package e2e
 
 import (
-	"bufio"
 	"context"
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
 	"strings"
 	"testing"
-
-	"github.com/onsi/ginkgo/reporters"
-
-	. "github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/config"
-	. "github.com/onsi/gomega"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
+	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/client"
 	"github.com/aws/aws-sdk-go/aws/session"
 	cfn "github.com/aws/aws-sdk-go/service/cloudformation"
-	ec2 "github.com/aws/aws-sdk-go/service/ec2"
+	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/iam"
-	awssts "github.com/aws/aws-sdk-go/service/sts"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/kubernetes"
-	infrav1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha3"
-	"sigs.k8s.io/cluster-api-provider-aws/cmd/clusterawsadm/cloudformation/bootstrap"
+	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo/config"
+	"github.com/onsi/ginkgo/reporters"
+	. "github.com/onsi/gomega"
+	"sigs.k8s.io/cluster-api-provider-aws/api/v1alpha3"
+	cfn_bootstrap "sigs.k8s.io/cluster-api-provider-aws/cmd/clusterawsadm/cloudformation/bootstrap"
 	cloudformation "sigs.k8s.io/cluster-api-provider-aws/cmd/clusterawsadm/cloudformation/service"
-	credentials "sigs.k8s.io/cluster-api-provider-aws/cmd/clusterawsadm/credentials"
+	"sigs.k8s.io/cluster-api-provider-aws/cmd/clusterawsadm/credentials"
 	"sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/awserrors"
-	sts "sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/services/sts"
-	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1alpha3"
-	common "sigs.k8s.io/cluster-api/test/helpers/components"
-	capiFlag "sigs.k8s.io/cluster-api/test/helpers/flag"
-	"sigs.k8s.io/cluster-api/test/helpers/kind"
-	"sigs.k8s.io/cluster-api/test/helpers/scheme"
-	"sigs.k8s.io/cluster-api/util"
-	crclient "sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/yaml"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"sigs.k8s.io/cluster-api-provider-aws/test/e2e/kubetest"
+	"sigs.k8s.io/cluster-api/test/framework"
+	"sigs.k8s.io/cluster-api/test/framework/bootstrap"
+	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 )
-
-func TestE2e(t *testing.T) {
-	RegisterFailHandler(Fail)
-
-	// If running in prow, make sure to output the junit files to the artifacts path
-	if ap, exists := os.LookupEnv("ARTIFACTS"); exists {
-		artifactPath = ap
-	}
-
-	junitPath := path.Join(artifactPath, fmt.Sprintf("junit.e2e_suite.%d.xml", config.GinkgoConfig.ParallelNode))
-	junitReporter := reporters.NewJUnitReporter(junitPath)
-
-	RunSpecsWithDefaultAndCustomReporters(t, "e2e Suite", []Reporter{junitReporter})
-}
 
 const (
-	capiNamespace       = "capi-system"
-	capiDeploymentName  = "capi-controller-manager"
-	cabpkNamespace      = "capi-kubeadm-bootstrap-system"
-	cabpkDeploymentName = "capi-kubeadm-bootstrap-controller-manager"
-	kcpNamespace        = "capi-kubeadm-control-plane-system"
-	kcpDeploymentName   = "capi-kubeadm-control-plane-controller-manager"
-	capaNamespace       = "capa-system"
-	capaDeploymentName  = "capa-controller-manager"
-	setupTimeout        = 10 * 60
-	stackName           = "cluster-api-provider-aws-sigs-k8s-io"
-	bootstrapUserName   = "bootstrapper.cluster-api-provider-aws.sigs.k8s.io"
-	keyPairName         = "cluster-api-provider-aws-sigs-k8s-io"
+	KubernetesVersion            = "KUBERNETES_VERSION"
+	CNIPath                      = "CNI"
+	AwsNodeMachineType           = "AWS_NODE_MACHINE_TYPE"
+	AwsAvailabilityZone1         = "AWS_AVAILABILITY_ZONE_1"
+	AwsAvailabilityZone2         = "AWS_AVAILABILITY_ZONE_2"
+	MultiAzFlavor                = "multi-az"
+	LimitAzFlavor                = "limit-az"
+	StorageClassFailureZoneLabel = "failure-domain.beta.kubernetes.io/zone"
 )
 
+// Test suite flags
 var (
-	managerImage    = capiFlag.DefineOrLookupStringFlag("managerImage", "", "Docker image to load into the kind cluster for testing")
-	capaComponents  = capiFlag.DefineOrLookupStringFlag("capaComponents", "", "capa components to load")
-	kustomizeBinary = capiFlag.DefineOrLookupStringFlag("kustomizeBinary", "kustomize", "path to the kustomize binary")
-	k8sVersion      = capiFlag.DefineOrLookupStringFlag("k8sVersion", "v1.17.3", "kubernetes version to test on")
-	sonobuoyVersion = capiFlag.DefineOrLookupStringFlag("sonobuoyVersion", "v0.17.2", "sonobuoy version")
+	// configPath is the path to the e2e config file.
+	configPath string
 
-	kindCluster       kind.Cluster
-	kindClient        crclient.Client
-	clientSet         *kubernetes.Clientset
-	sess              client.ConfigProvider
-	accountID         string
-	accessKey         *iam.AccessKey
-	suiteTmpDir       string
-	region            string
-	artifactPath      = ".artifacts"
-	logPath           string
-	bootstrapTemplate bootstrap.Template
+	// useExistingCluster instructs the test to use the current cluster instead of creating a new one (default discovery rules apply).
+	useExistingCluster bool
+
+	// artifactFolder is the folder to store e2e test artifacts.
+	artifactFolder string
+
+	// skipCleanup prevents cleanup of test resources e.g. for debug purposes.
+	skipCleanup bool
+
+	// skipCloudFormationDeletion prevents the deletion of the AWS CloudFormation stack
+	skipCloudFormationDeletion bool
 )
 
-func createBootstrapTemplate() bootstrap.Template {
-	t := bootstrap.NewTemplate()
-	t.Spec.BootstrapUser.Enable = true
-	t.Spec.BootstrapUser.UserName = bootstrapUserName
-	t.Spec.StackName = stackName
-	t.Spec.Region = region
-	return t
+// Test suite global vars
+var (
+	// e2eConfig to be used for this test, read from configPath.
+	e2eConfig *clusterctl.E2EConfig
+
+	// clusterctlConfigPath to be used for this test, created by generating a clusterctl local repository
+	// with the providers specified in the configPath.
+	clusterctlConfigPath string
+
+	// bootstrapClusterProvider manages provisioning of the the bootstrap cluster to be used for the e2e tests.
+	// Please note that provisioning will be skipped if use-existing-cluster is provided.
+	bootstrapClusterProvider bootstrap.ClusterProvider
+
+	// bootstrapClusterProxy allows to interact with the bootstrap cluster to be used for the e2e tests.
+	bootstrapClusterProxy framework.ClusterProxy
+
+	// bootstrapTemplate is the clusterawsadm bootstrap template for this run
+	bootstrapTemplate *cfn_bootstrap.Template
+
+	// bootstrapAccessKey
+	bootstrapAccessKey *iam.AccessKey
+
+	defaultSSHKeyPairName = "cluster-api-provider-aws-sigs-k8s-io"
+
+	// conformance configuration
+	kubetestConfig *kubetest.Configuration
+
+	// ticker for dumping resources
+	resourceTicker *time.Ticker
+
+	// tickerDone to stop ticking
+	resourceTickerDone chan bool
+
+	// ticker for dumping resources
+	machineTicker *time.Ticker
+
+	// tickerDone to stop ticking
+	machineTickerDone chan bool
+)
+
+func init() {
+	flag.StringVar(&configPath, "config-path", "", "path to the e2e config file")
+	flag.StringVar(&artifactFolder, "artifacts-folder", "", "folder where e2e test artifact should be stored")
+	flag.BoolVar(&useExistingCluster, "use-existing-cluster", false, "if true, the test uses the current cluster instead of creating a new one (default discovery rules apply)")
+	flag.BoolVar(&skipCleanup, "skip-cleanup", false, "if true, the resource cleanup after tests will be skipped")
+	flag.BoolVar(&skipCloudFormationDeletion, "skip-cloudformation-deletion", false, "if true, an AWS CloudFormation stack will not be deleted")
+}
+
+type synchronizedBeforeTestSuiteConfig struct {
+	ArtifactFolder       string                 `json:"artifactFolder,omitempty"`
+	ConfigPath           string                 `json:"configPath,omitempty"`
+	ClusterctlConfigPath string                 `json:"clusterctlConfigPath,omitempty"`
+	KubeconfigPath       string                 `json:"kubeconfigPath,omitempty"`
+	Region               string                 `json:"region,omitempty"`
+	E2EConfig            clusterctl.E2EConfig   `json:"e2eConfig,omitempty"`
+	KubetestConfig       kubetest.Configuration `json:"conformanceConfiguration,omitempty"`
+}
+
+func TestE2E(t *testing.T) {
+	RegisterFailHandler(Fail)
+	junitPath := path.Join(artifactFolder, fmt.Sprintf("junit.e2e_suite.%d.xml", config.GinkgoConfig.ParallelNode))
+	junitReporter := reporters.NewJUnitReporter(junitPath)
+
+	RunSpecsWithDefaultAndCustomReporters(t, "capa-e2e", []Reporter{junitReporter})
 }
 
 var _ = SynchronizedBeforeSuite(func() []byte {
-	artifactPath, _ = os.LookupEnv("ARTIFACTS")
-	logPath = path.Join(artifactPath, "logs")
-	Expect(os.MkdirAll(filepath.Dir(logPath), 0o755)).To(Succeed())
+	flag.Parse()
+	Expect(configPath).To(BeAnExistingFile(), "Invalid test suite argument. configPath should be an existing file.")
+	Expect(os.MkdirAll(artifactFolder, 0o750)).To(Succeed(), "Invalid test suite argument. Can't create artifacts-folder %q", artifactFolder)
+	Byf("Loading the e2e test configuration from %q", configPath)
+	e2eConfig = loadE2EConfig(configPath)
+	Expect(prepConformanceConfig(artifactFolder, e2eConfig)).To(Succeed())
+	awsSession = newAWSSession()
+	createCloudFormationStack(awsSession, getBootstrapTemplate())
+	ensureNoServiceLinkedRoles(awsSession)
+	ensureSSHKeyPair(awsSession, defaultSSHKeyPairName)
+	bootstrapAccessKey = newUserAccessKey(awsSession, getBootstrapTemplate().Spec.BootstrapUser.UserName)
 
-	fmt.Fprintf(GinkgoWriter, "Setting up kind cluster\n")
+	By("Initializing a runtime.Scheme with all the GVK relevant for this test")
+	scheme := initScheme()
 
-	fmt.Fprintf(GinkgoWriter, "Creating AWS prerequisites\n")
-	sess = getSession()
-	accountID = getAccountID(sess)
-	var ok bool
-	region, ok = os.LookupEnv("AWS_REGION")
-	fmt.Fprintf(GinkgoWriter, "Running in region: %s\n", region)
-	if !ok {
-		fmt.Fprintf(GinkgoWriter, "Environment variable AWS_REGION not found")
-		Expect(ok).To(BeTrue())
+	kubetestConfiguration := kubetest.NewConfiguration(e2eConfig, artifactFolder)
+	kubetestConfig = kubetestConfiguration
+	// If using a version of Kubernetes from CI, override the image ID with a known good image
+	if kubetestConfig.UseCIArtifacts {
+		e2eConfig.Variables["IMAGE_ID"] = conformanceImageID()
 	}
-	bootstrapTemplate = createBootstrapTemplate()
-	createKeyPair(sess)
-	createIAMRoles(sess, bootstrapTemplate)
 
-	iamc := iam.New(sess)
-	out, err := iamc.CreateAccessKey(&iam.CreateAccessKeyInput{UserName: aws.String(bootstrapTemplate.Spec.BootstrapUser.UserName)})
+	Byf("Creating a clusterctl local repository into %q", artifactFolder)
+	clusterctlConfigPath = createClusterctlLocalRepository(e2eConfig, filepath.Join(artifactFolder, "repository"))
+
+	By("Setting up the bootstrap cluster")
+	bootstrapClusterProvider, bootstrapClusterProxy = setupBootstrapCluster(e2eConfig, scheme, useExistingCluster)
+
+	setEnvVar("AWS_B64ENCODED_CREDENTIALS", encodeCredentials(bootstrapAccessKey, getBootstrapTemplate().Spec.Region), true)
+
+	By("Initializing the bootstrap cluster")
+	initBootstrapCluster(bootstrapClusterProxy, e2eConfig, clusterctlConfigPath, artifactFolder)
+
+	conf := synchronizedBeforeTestSuiteConfig{
+		ArtifactFolder:       artifactFolder,
+		ConfigPath:           configPath,
+		ClusterctlConfigPath: clusterctlConfigPath,
+		KubeconfigPath:       bootstrapClusterProxy.GetKubeconfigPath(),
+		Region:               getBootstrapTemplate().Spec.Region,
+		E2EConfig:            *e2eConfig,
+		KubetestConfig:       *kubetestConfiguration,
+	}
+
+	data, err := yaml.Marshal(conf)
 	Expect(err).NotTo(HaveOccurred())
-	Expect(out.AccessKey).NotTo(BeNil())
-	return []byte(
-		strings.Join(
-			[]string{
-				aws.StringValue(out.AccessKey.UserName),
-				aws.StringValue(out.AccessKey.AccessKeyId),
-				aws.StringValue(out.AccessKey.SecretAccessKey),
-			},
-			",",
-		),
-	)
-}, func(accessKeyPair []byte) {
-	parts := strings.Split(string(accessKeyPair), ",")
-	Expect(parts).To(HaveLen(3))
+	return data
 
-	accessKeyUsername := parts[0]
-	accessKeyID := parts[1]
-	secretAccessKey := parts[2]
-
-	accessKey = &iam.AccessKey{
-		AccessKeyId:     &accessKeyID,
-		SecretAccessKey: &secretAccessKey,
-		UserName:        &accessKeyUsername,
-	}
-	var ok bool
-	region, ok = os.LookupEnv("AWS_REGION")
-	fmt.Fprintf(GinkgoWriter, "Running in region: %s\n", region)
-	if !ok {
-		fmt.Fprintf(GinkgoWriter, "Environment variable AWS_REGION not found")
-		Expect(ok).To(BeTrue())
-	}
-	fmt.Fprintf(GinkgoWriter, "GETTING SESSION")
-	sess = getSession()
-	fmt.Fprintf(GinkgoWriter, "... DONE GETTING SESSION")
-
-	var err error
-
-	suiteTmpDir, err = ioutil.TempDir("", "capa-e2e-suite")
+}, func(data []byte) {
+	// Before each ParallelNode.
+	conf := &synchronizedBeforeTestSuiteConfig{}
+	err := yaml.UnmarshalStrict(data, conf)
 	Expect(err).NotTo(HaveOccurred())
+	artifactFolder = conf.ArtifactFolder
+	configPath = conf.ConfigPath
+	clusterctlConfigPath = conf.ClusterctlConfigPath
+	kubetestConfig = &conf.KubetestConfig
+	bootstrapClusterProxy = framework.NewClusterProxy("bootstrap", conf.KubeconfigPath, initScheme())
+	e2eConfig = &conf.E2EConfig
+	azs := getAvailabilityZones()
+	setEnvVar(AwsAvailabilityZone1, *azs[0].ZoneName, false)
+	setEnvVar(AwsAvailabilityZone2, *azs[1].ZoneName, false)
+	setEnvVar("AWS_REGION", conf.Region, false)
+	setEnvVar("AWS_SSH_KEY_NAME", defaultSSHKeyPairName, false)
+	awsSession = newAWSSession()
+	namespaces = map[*corev1.Namespace]context.CancelFunc{}
+	resourceTicker = time.NewTicker(time.Second * 5)
+	resourceTickerDone = make(chan bool)
+	// Get EC2 logs every minute
+	machineTicker = time.NewTicker(time.Second * 60)
+	machineTickerDone = make(chan bool)
+	resourceCtx, resourceCancel := context.WithCancel(context.Background())
+	machineCtx, machineCancel := context.WithCancel(context.Background())
 
-	kindCluster = kind.Cluster{
-		Name: "capa-test-" + util.RandomString(6),
-	}
-	kindCluster.Setup()
-	loadManagerImage(kindCluster)
+	// Dump resources every 5 seconds
+	go func() {
+		defer GinkgoRecover()
+		for {
+			select {
+			case <-resourceTickerDone:
+				resourceCancel()
+				return
+			case <-resourceTicker.C:
+				for k := range namespaces {
+					dumpSpecResources(resourceCtx, bootstrapClusterProxy, artifactFolder, k)
+				}
+			}
+		}
+	}()
 
-	// create the management cluster clients we'll need
-	restConfig := kindCluster.RestConfig()
-	mapper, err := apiutil.NewDynamicRESTMapper(restConfig, apiutil.WithLazyDiscovery)
-	Expect(err).NotTo(HaveOccurred())
-	kindClient, err = crclient.New(kindCluster.RestConfig(), crclient.Options{Scheme: setupScheme(), Mapper: mapper})
-	Expect(err).NotTo(HaveOccurred())
-	clientSet, err = kubernetes.NewForConfig(kindCluster.RestConfig())
-	Expect(err).NotTo(HaveOccurred())
-
-	// Deploy CertManager
-	certmanagerYaml := "https://github.com/jetstack/cert-manager/releases/download/v0.11.0/cert-manager.yaml"
-	applyManifests(kindCluster, &certmanagerYaml)
-
-	// Wait for CertManager to be available before continuing
-	common.WaitDeployment(kindClient, "cert-manager", "cert-manager-webhook")
-
-	// Deploy the CAPI, CABPK, and KCP components from Cluster API repository,
-	capiYAMLssss := "https://github.com/kubernetes-sigs/cluster-api/releases/download/v0.3.6/cluster-api-components.yaml"
-	applyManifests(kindCluster, &capiYAMLssss)
-
-	// Deploy the CAPA components
-	deployCAPAComponents(kindCluster)
-
-	// Verify capi components are deployed
-	common.WaitDeployment(kindClient, capiNamespace, capiDeploymentName)
-	watchLogs(capiNamespace, capiDeploymentName, logPath)
-
-	// Verify cabpk components are deployed
-	common.WaitDeployment(kindClient, cabpkNamespace, cabpkDeploymentName)
-	watchLogs(cabpkNamespace, cabpkDeploymentName, logPath)
-
-	// Verify kcp components are deployed
-	common.WaitDeployment(kindClient, kcpNamespace, kcpDeploymentName)
-	watchLogs(kcpNamespace, kcpDeploymentName, logPath)
-
-	// Verify capa components are deployed
-	common.WaitDeployment(kindClient, capaNamespace, capaDeploymentName)
-	watchLogs(capaNamespace, capaDeploymentName, logPath)
-}, setupTimeout)
-
-var _ = SynchronizedAfterSuite(func() {
-	fmt.Fprintf(GinkgoWriter, "Tearing down kind cluster\n")
-
-	if kindCluster.Name != "" {
-		kindCluster.Teardown()
-	}
-	if suiteTmpDir != "" {
-		os.RemoveAll(suiteTmpDir)
-	}
-}, func() {
-	// This is intentionally done per node
-	iamc := iam.New(sess)
-	iamc.DeleteAccessKey(&iam.DeleteAccessKeyInput{UserName: accessKey.UserName, AccessKeyId: accessKey.AccessKeyId})
-	deleteIAMRoles(sess, bootstrapTemplate)
+	// Dump machine logs every 60 seconds
+	go func() {
+		defer GinkgoRecover()
+		for {
+			select {
+			case <-machineTickerDone:
+				machineCancel()
+				return
+			case <-machineTicker.C:
+				for k := range namespaces {
+					dumpMachines(machineCtx, bootstrapClusterProxy, k, artifactFolder)
+				}
+			}
+		}
+	}()
 })
 
-// watchLogs streams logs for all containers for all pods belonging to a deployment. Each container's logs are streamed
-// in a separate goroutine so they can all be streamed concurrently. This only causes a test failure if there are errors
-// retrieving the deployment, its pods, or setting up a log file. If there is an error with the log streaming itself,
-// that does not cause the test to fail.
-func watchLogs(namespace, deploymentName, logDir string) {
-	deployment := &appsv1.Deployment{}
-	Expect(kindClient.Get(context.TODO(), crclient.ObjectKey{Namespace: namespace, Name: deploymentName}, deployment)).To(Succeed())
-
-	selector, err := metav1.LabelSelectorAsMap(deployment.Spec.Selector)
-	Expect(err).NotTo(HaveOccurred())
-
-	pods := &corev1.PodList{}
-	Expect(kindClient.List(context.TODO(), pods, crclient.InNamespace(namespace), crclient.MatchingLabels(selector))).To(Succeed())
-
-	for _, pod := range pods.Items {
-		for _, container := range deployment.Spec.Template.Spec.Containers {
-			// Watch each container's logs in a goroutine so we can stream them all concurrently.
-			go func(pod corev1.Pod, container corev1.Container) {
-				defer GinkgoRecover()
-
-				logFile := path.Join(logDir, deploymentName, pod.Name, container.Name+".log")
-				fmt.Fprintf(GinkgoWriter, "Creating directory: %s\n", filepath.Dir(logFile))
-				Expect(os.MkdirAll(filepath.Dir(logFile), 0o755)).To(Succeed())
-
-				fmt.Fprintf(GinkgoWriter, "Creating file: %s\n", logFile)
-				f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644)
-				Expect(err).NotTo(HaveOccurred())
-				defer f.Close()
-
-				opts := &corev1.PodLogOptions{
-					Container: container.Name,
-					Follow:    true,
-				}
-
-				podLogs, err := clientSet.CoreV1().Pods(namespace).GetLogs(pod.Name, opts).Stream()
-				if err != nil {
-					// Failing to stream logs should not cause the test to fail
-					fmt.Fprintf(GinkgoWriter, "Error starting logs stream for pod %s/%s, container %s: %v\n", namespace, pod.Name, container.Name, err)
-					return
-				}
-				defer podLogs.Close()
-
-				out := bufio.NewWriter(f)
-				defer out.Flush()
-				_, err = out.ReadFrom(podLogs)
-				if err != nil && err.Error() != "unexpected EOF" {
-					// Failing to stream logs should not cause the test to fail
-					fmt.Fprintf(GinkgoWriter, "Got error while streaming logs for pod %s/%s, container %s: %v\n", namespace, pod.Name, container.Name, err)
-				}
-			}(pod, container)
+// Using a SynchronizedAfterSuite for controlling how to delete resources shared across ParallelNodes (~ginkgo threads).
+// The bootstrap cluster is shared across all the tests, so it should be deleted only after all ParallelNodes completes.
+// The local clusterctl repository is preserved like everything else created into the artifact folder.
+var _ = SynchronizedAfterSuite(func() {
+	if resourceTickerDone != nil {
+		resourceTickerDone <- true
+	}
+	if machineTickerDone != nil {
+		machineTickerDone <- true
+	}
+	kubetest.GatherReports(kubetestConfig)
+	ctx, cancel := context.WithTimeout(context.TODO(), 5*time.Minute)
+	defer cancel()
+	for k := range namespaces {
+		dumpSpecResourcesAndCleanup(ctx, "", bootstrapClusterProxy, artifactFolder, k, e2eConfig.GetIntervals, skipCleanup)
+		dumpMachines(ctx, bootstrapClusterProxy, k, artifactFolder)
+	}
+}, func() {
+	// After all ParallelNodes.
+	By("Tearing down the management cluster")
+	if !skipCleanup {
+		tearDown(bootstrapClusterProvider, bootstrapClusterProxy)
+		if !skipCloudFormationDeletion {
+			awsSession = newAWSSession()
+			deleteCloudFormationStack(awsSession, getBootstrapTemplate())
 		}
+	}
+})
+
+// initScheme creates a new GVK scheme
+func initScheme() *runtime.Scheme {
+	sc := runtime.NewScheme()
+	framework.TryAddDefaultSchemes(sc)
+	_ = v1alpha3.AddToScheme(sc)
+	_ = clientgoscheme.AddToScheme(sc)
+	return sc
+}
+
+// loadE2EConfig loads the e2econfig from the specified path
+func loadE2EConfig(configPath string) *clusterctl.E2EConfig {
+	config := clusterctl.LoadE2EConfig(context.TODO(), clusterctl.LoadE2EConfigInput{ConfigPath: configPath})
+	Expect(config).ToNot(BeNil(), "Failed to load E2E config from %s", configPath)
+	return config
+}
+
+// createClusterctlLocalRepository generates a clusterctl repository.
+// Must always be run after kubetest.NewConfiguration
+func createClusterctlLocalRepository(config *clusterctl.E2EConfig, repositoryFolder string) string {
+	clusterctlConfig := clusterctl.CreateRepository(context.TODO(), clusterctl.CreateRepositoryInput{
+		E2EConfig:        config,
+		RepositoryFolder: repositoryFolder,
+	})
+	Expect(clusterctlConfig).To(BeAnExistingFile(), "The clusterctl config file does not exists in the local repository %s", repositoryFolder)
+	return clusterctlConfig
+}
+
+// setupBootstrapCluster installs Cluster API components via clusterctl
+func setupBootstrapCluster(config *clusterctl.E2EConfig, scheme *runtime.Scheme, useExistingCluster bool) (bootstrap.ClusterProvider, framework.ClusterProxy) {
+	var clusterProvider bootstrap.ClusterProvider
+	kubeconfigPath := ""
+	if !useExistingCluster {
+		clusterProvider = bootstrap.CreateKindBootstrapClusterAndLoadImages(context.TODO(), bootstrap.CreateKindBootstrapClusterAndLoadImagesInput{
+			Name:               config.ManagementClusterName,
+			RequiresDockerSock: config.HasDockerProvider(),
+			Images:             config.Images,
+		})
+		Expect(clusterProvider).ToNot(BeNil(), "Failed to create a bootstrap cluster")
+
+		kubeconfigPath = clusterProvider.GetKubeconfigPath()
+		Expect(kubeconfigPath).To(BeAnExistingFile(), "Failed to get the kubeconfig file for the bootstrap cluster")
+	}
+
+	clusterProxy := framework.NewClusterProxy("bootstrap", kubeconfigPath, scheme)
+	Expect(clusterProxy).ToNot(BeNil(), "Failed to get a bootstrap cluster proxy")
+
+	return clusterProvider, clusterProxy
+}
+
+// initBootstrapCluster uses kind to create a cluster
+func initBootstrapCluster(bootstrapClusterProxy framework.ClusterProxy, config *clusterctl.E2EConfig, clusterctlConfig, artifactFolder string) {
+	clusterctl.InitManagementClusterAndWatchControllerLogs(context.TODO(), clusterctl.InitManagementClusterAndWatchControllerLogsInput{
+		ClusterProxy:            bootstrapClusterProxy,
+		ClusterctlConfigPath:    clusterctlConfig,
+		InfrastructureProviders: config.InfrastructureProviders(),
+		LogFolder:               filepath.Join(artifactFolder, "clusters", bootstrapClusterProxy.GetName()),
+	}, config.GetIntervals(bootstrapClusterProxy.GetName(), "wait-controllers")...)
+}
+
+// tearDown the bootstrap kind cluster
+func tearDown(bootstrapClusterProvider bootstrap.ClusterProvider, bootstrapClusterProxy framework.ClusterProxy) {
+	if bootstrapClusterProxy != nil {
+		bootstrapClusterProxy.Dispose(context.TODO())
+	}
+	if bootstrapClusterProvider != nil {
+		bootstrapClusterProvider.Dispose(context.TODO())
 	}
 }
 
-func getSession() client.ConfigProvider {
+// newBootstrapTemplate generates a clusterawsadm configuration, and prints it
+// and the resultant cloudformation template to the artifacts directory
+func newBootstrapTemplate() *cfn_bootstrap.Template {
+	By("Creating a bootstrap AWSIAMConfiguration")
+	t := cfn_bootstrap.NewTemplate()
+	t.Spec.BootstrapUser.Enable = true
+	region, err := credentials.ResolveRegion("")
+	Expect(err).NotTo(HaveOccurred())
+	t.Spec.Region = region
+	str, err := yaml.Marshal(t.Spec)
+	Expect(err).NotTo(HaveOccurred())
+	Expect(ioutil.WriteFile(path.Join(artifactFolder, "awsiamconfiguration.yaml"), str, 0644)).To(Succeed())
+	cfnData, err := t.RenderCloudFormation().YAML()
+	Expect(err).NotTo(HaveOccurred())
+	Expect(ioutil.WriteFile(path.Join(artifactFolder, "cloudformation.yaml"), cfnData, 0644)).To(Succeed())
+	return &t
+}
+
+// getBootstrapTemplate gets or generates a new bootstrap template
+func getBootstrapTemplate() *cfn_bootstrap.Template {
+	if bootstrapTemplate == nil {
+		bootstrapTemplate = newBootstrapTemplate()
+	}
+	return bootstrapTemplate
+}
+
+func newAWSSession() client.ConfigProvider {
+	By("Getting an AWS IAM session")
+	region, err := credentials.ResolveRegion("")
+	Expect(err).NotTo(HaveOccurred())
+	config := aws.NewConfig().WithCredentialsChainVerboseErrors(true).WithRegion(region)
 	sess, err := session.NewSessionWithOptions(session.Options{
 		SharedConfigState: session.SharedConfigEnable,
+		Config:            *config,
 	})
+	Expect(err).NotTo(HaveOccurred())
+	_, err = sess.Config.Credentials.Get()
 	Expect(err).NotTo(HaveOccurred())
 	return sess
 }
 
-func getAccountID(prov client.ConfigProvider) string {
-	if prov == nil {
-		panic("The client configuration provider for AWS Cannot be nil !!!")
+// ensureNotServiceLinkedRoles removes an auto-created IAM role, and tests
+// the controller's IAM permissions to use ELB successfully
+func ensureNoServiceLinkedRoles(prov client.ConfigProvider) {
+	Byf("Deleting AWS IAM Service Linked Role: role-name=AWSServiceRoleForElasticLoadBalancing")
+	iamSvc := iam.New(prov)
+	_, err := iamSvc.DeleteServiceLinkedRole(&iam.DeleteServiceLinkedRoleInput{
+		RoleName: aws.String("AWSServiceRoleForElasticLoadBalancing"),
+	})
+	if code, _ := awserrors.Code(err); code != iam.ErrCodeNoSuchEntityException {
+		Expect(err).NotTo(HaveOccurred())
 	}
-	stsSvc := sts.NewService(awssts.New(prov))
-	accountID, err := stsSvc.AccountID()
-	Expect(err).NotTo(HaveOccurred())
-	return accountID
 }
 
-func createIAMRoles(prov client.ConfigProvider, t bootstrap.Template) {
+// createCloudFormationStack ensures the cloudformation stack is up to date
+func createCloudFormationStack(prov client.ConfigProvider, t *cfn_bootstrap.Template) {
+	Byf("Creating AWS CloudFormation stack for AWS IAM resources: stack-name=%s", t.Spec.StackName)
 	cfnSvc := cloudformation.NewService(cfn.New(prov))
 	cfnTemplate := t.RenderCloudFormation()
 	Expect(
@@ -328,14 +420,18 @@ func createIAMRoles(prov client.ConfigProvider, t bootstrap.Template) {
 	).To(Succeed())
 }
 
-func deleteIAMRoles(prov client.ConfigProvider, t bootstrap.Template) {
+// deleteCloudFormationStack removes the provisioned clusterawsadm stack
+func deleteCloudFormationStack(prov client.ConfigProvider, t *cfn_bootstrap.Template) {
+	Byf("Deleting %s CloudFormation stack", t.Spec.StackName)
 	cfnSvc := cloudformation.NewService(cfn.New(prov))
 	Expect(
 		cfnSvc.DeleteStack(t.Spec.StackName),
 	).To(Succeed())
 }
 
-func createKeyPair(prov client.ConfigProvider) {
+// ensureSSHKeyPair ensures A SSH key is present under the name
+func ensureSSHKeyPair(prov client.ConfigProvider, keyPairName string) {
+	Byf("Ensuring presence of SSH key in EC2: key-name=%s", keyPairName)
 	ec2c := ec2.New(prov)
 	_, err := ec2c.CreateKeyPair(&ec2.CreateKeyPairInput{KeyName: aws.String(keyPairName)})
 	if code, _ := awserrors.Code(err); code != "InvalidKeyPair.Duplicate" {
@@ -343,67 +439,81 @@ func createKeyPair(prov client.ConfigProvider) {
 	}
 }
 
-func loadManagerImage(kindCluster kind.Cluster) {
-	if managerImage != nil && *managerImage != "" {
-		kindCluster.LoadImage(*managerImage)
-	}
-}
-
-func applyManifests(kindCluster kind.Cluster, manifests *string) {
-	Expect(manifests).ToNot(BeNil())
-	fmt.Fprintf(GinkgoWriter, "Applying manifests for %s\n", *manifests)
-	Expect(*manifests).ToNot(BeEmpty())
-	kindCluster.ApplyYAML(*manifests)
-}
-
-func deployCAPAComponents(kindCluster kind.Cluster) {
-	if capaComponents != nil && *capaComponents != "" {
-		applyManifests(kindCluster, capaComponents)
-		return
-	}
-
-	fmt.Fprintf(GinkgoWriter, "Generating CAPA manifests\n")
-
-	// Build the manifests using kustomize
-	capaManifests, err := exec.Command(*kustomizeBinary, "build", "../../config").Output()
-	if err != nil {
-		if exitError, ok := err.(*exec.ExitError); ok {
-			fmt.Fprintf(GinkgoWriter, "Error: %s\n", string(exitError.Stderr))
-		}
-	}
-	Expect(err).NotTo(HaveOccurred())
-
-	// envsubst the credentials
-	Expect(err).NotTo(HaveOccurred())
-	b64credentials := generateB64Credentials()
-	os.Setenv("AWS_B64ENCODED_CREDENTIALS", b64credentials)
-	manifestsContent := os.ExpandEnv(string(capaManifests))
-
-	// write out the manifests
-	manifestFile := path.Join(suiteTmpDir, "infrastructure-components.yaml")
-	Expect(ioutil.WriteFile(manifestFile, []byte(manifestsContent), 0o644)).To(Succeed())
-
-	// apply generated manifests
-	applyManifests(kindCluster, &manifestFile)
-}
-
-func generateB64Credentials() string {
+// encodeCredentials leverages clusterawsadm to encode AWS credentials
+func encodeCredentials(accessKey *iam.AccessKey, region string) string {
 	creds := credentials.AWSCredentials{
 		Region:          region,
 		AccessKeyID:     *accessKey.AccessKeyId,
 		SecretAccessKey: *accessKey.SecretAccessKey,
 	}
-
 	encCreds, err := creds.RenderBase64EncodedAWSDefaultProfile()
-
 	Expect(err).NotTo(HaveOccurred())
-
 	return encCreds
 }
 
-func setupScheme() *runtime.Scheme {
-	s := scheme.SetupScheme()
-	Expect(bootstrapv1.AddToScheme(s)).To(Succeed())
-	Expect(infrav1.AddToScheme(s)).To(Succeed())
-	return s
+// setEnvVar sets an environment variable in the process. If marked private,
+// the value is not printed.
+func setEnvVar(key, value string, private bool) {
+	printableValue := "*******"
+	if !private {
+		printableValue = value
+	}
+
+	Byf("Setting environment variable: key=%s, value=%s", key, printableValue)
+	os.Setenv(key, value)
+}
+
+// newUserAccessKey generates a new AWS Access Key pair based off of the
+// bootstrap user. This tests that the CloudFormation policy is correct.
+func newUserAccessKey(prov client.ConfigProvider, userName string) *iam.AccessKey {
+	iamSvc := iam.New(prov)
+	keyOuts, _ := iamSvc.ListAccessKeys(&iam.ListAccessKeysInput{
+		UserName: aws.String(userName),
+	})
+	for i := range keyOuts.AccessKeyMetadata {
+		Byf("Deleting an existing access key: user-name=%s", userName)
+		_, err := iamSvc.DeleteAccessKey(&iam.DeleteAccessKeyInput{
+			UserName:    aws.String(userName),
+			AccessKeyId: keyOuts.AccessKeyMetadata[i].AccessKeyId,
+		})
+		Expect(err).NotTo(HaveOccurred())
+	}
+	Byf("Creating an access key: user-name=%s", userName)
+	out, err := iamSvc.CreateAccessKey(&iam.CreateAccessKeyInput{UserName: aws.String(userName)})
+	Expect(err).NotTo(HaveOccurred())
+	Expect(out.AccessKey).ToNot(BeNil())
+
+	return &iam.AccessKey{
+		AccessKeyId:     out.AccessKey.AccessKeyId,
+		SecretAccessKey: out.AccessKey.SecretAccessKey,
+	}
+}
+
+// conformanceImageID looks up a specific image for a given
+// Kubernetes version in the e2econfig
+func conformanceImageID() string {
+	ver := e2eConfig.GetVariable("CONFORMANCE_CI_ARTIFACTS_KUBERNETES_VERSION")
+	strippedVer := strings.Replace(ver, "v", "", 1)
+	amiName := AMIPrefix + strippedVer + "*"
+
+	Byf("Searching for AMI: name=%s", amiName)
+	ec2Svc := ec2.New(awsSession)
+	filters := []*ec2.Filter{
+		{
+			Name:   aws.String("name"),
+			Values: []*string{aws.String(amiName)},
+		},
+	}
+	filters = append(filters, &ec2.Filter{
+		Name:   aws.String("owner-id"),
+		Values: []*string{aws.String(DefaultImageLookupOrg)},
+	})
+	resp, err := ec2Svc.DescribeImages(&ec2.DescribeImagesInput{
+		Filters: filters,
+	})
+	Expect(err).NotTo(HaveOccurred())
+	Expect(len(resp.Images)).To(Not(BeZero()))
+	imageID := aws.StringValue(resp.Images[0].ImageId)
+	Byf("Using AMI: image-id=%s", imageID)
+	return imageID
 }
