+++ test/e2e/exec.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_new
+package e2e
 
 import (
 	"context"
