--- test/e2e/go.mod
@@ -1,27 +1,0 @@
-module sigs.k8s.io/cluster-api-provider-aws/test/e2e
-
-go 1.13
-
-require (
-	github.com/aws/aws-sdk-go v1.34.10
-	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
-	github.com/gophercloud/gophercloud v0.2.0
-	github.com/onsi/ginkgo v1.12.2
-	github.com/onsi/gomega v1.10.1
-	github.com/pkg/errors v0.9.1
-	github.com/vmware-tanzu/sonobuoy v0.18.2
-	golang.org/x/sync v0.0.0-20200317015054-43a5402ce75a
-	golang.org/x/time v0.0.0-20200416051211-89c76fbcd5d1 // indirect
-	k8s.io/api v0.17.8
-	k8s.io/apimachinery v0.17.8
-	k8s.io/client-go v11.0.1-0.20190704100234-640d9f240853+incompatible
-	k8s.io/utils v0.0.0-20200619165400-6e3d28b6ed19
-	sigs.k8s.io/cluster-api v0.3.7
-	sigs.k8s.io/cluster-api-provider-aws v0.5.3
-	sigs.k8s.io/controller-runtime v0.5.8
-)
-
-replace (
-	k8s.io/client-go => k8s.io/client-go v0.0.0-20190918160344-1fbdaa4c8d90
-	sigs.k8s.io/cluster-api-provider-aws => ../../
-)
