--- test/e2e/sonobuoy-config-tmpl.yaml
@@ -1,183 +1,0 @@
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: sonobuoy
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  labels:
-    component: sonobuoy
-  name: sonobuoy-serviceaccount
-  namespace: sonobuoy
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  labels:
-    component: sonobuoy
-  name: sonobuoy-serviceaccount-sonobuoy
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: sonobuoy-serviceaccount
-subjects:
-- kind: ServiceAccount
-  name: sonobuoy-serviceaccount
-  namespace: sonobuoy
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  labels:
-    component: sonobuoy
-  name: sonobuoy-serviceaccount
-rules:
-- apiGroups:
-  - '*'
-  resources:
-  - '*'
-  verbs:
-  - '*'
-- nonResourceURLs:
-  - '/metrics'
-  - '/logs'
-  - '/logs/*'
-  verbs:
-  - 'get'
----
-apiVersion: v1
-data:
-  config.json: |
-    {"Description":"DEFAULT","Version":"{{.SonobuoyVersion}}","ResultsDir":"/tmp/sonobuoy","Resources":["apiservices","certificatesigningrequests","clusterrolebindings","clusterroles","componentstatuses","configmaps","controllerrevisions","cronjobs","customresourcedefinitions","daemonsets","deployments","endpoints","ingresses","jobs","leases","limitranges","mutatingwebhookconfigurations","namespaces","networkpolicies","nodes","persistentvolumeclaims","persistentvolumes","poddisruptionbudgets","pods","podlogs","podsecuritypolicies","podtemplates","priorityclasses","replicasets","replicationcontrollers","resourcequotas","rolebindings","roles","servergroups","serverversion","serviceaccounts","services","statefulsets","storageclasses","validatingwebhookconfigurations","volumeattachments"],"Filters":{"Namespaces":".*","LabelSelector":""},"Limits":{"PodLogs":{"Namespaces":"","SonobuoyNamespace":true,"FieldSelectors":[],"LabelSelector":"","Previous":false,"SinceSeconds":null,"SinceTime":null,"Timestamps":false,"TailLines":null,"LimitBytes":null,"LimitSize":"","LimitTime":""}},"QPS":30,"Burst":50,"Server":{"bindaddress":"0.0.0.0","bindport":8080,"advertiseaddress":"","timeoutseconds":10800},"Plugins":[{"name":"e2e"},{"name":"systemd-logs"}],"PluginSearchPath":["./plugins.d","/etc/sonobuoy/plugins.d","~/sonobuoy/plugins.d"],"Namespace":"sonobuoy","WorkerImage":"sonobuoy/sonobuoy:{{.SonobuoyVersion}}","ImagePullPolicy":"IfNotPresent","ImagePullSecrets":"","ProgressUpdatesPort":"8099"}
-kind: ConfigMap
-metadata:
-  labels:
-    component: sonobuoy
-  name: sonobuoy-config-cm
-  namespace: sonobuoy
----
-apiVersion: v1
-data:
-  plugin-0.yaml: |
-    sonobuoy-config:
-      driver: Job
-      plugin-name: e2e
-      result-format: junit
-    spec:
-      command:
-      - /run_e2e.sh
-      env:
-      - name: E2E_FOCUS
-        value: \[Conformance\]
-      - name: E2E_SKIP
-      - name: E2E_PARALLEL
-        value: "1"
-      - name: E2E_USE_GO_RUNNER
-        value: "true"
-      image: gcr.io/google-containers/conformance:{{.K8sVersion}}
-      imagePullPolicy: Always
-      name: e2e
-      resources: {}
-      volumeMounts:
-      - mountPath: /tmp/results
-        name: results
-  plugin-1.yaml: |
-    sonobuoy-config:
-      driver: DaemonSet
-      plugin-name: systemd-logs
-      result-format: raw
-    spec:
-      command:
-      - /bin/sh
-      - -c
-      - /get_systemd_logs.sh && sleep 3600
-      env:
-      - name: CHROOT_DIR
-        value: /node
-      - name: RESULTS_DIR
-        value: /tmp/results
-      - name: NODE_NAME
-        valueFrom:
-          fieldRef:
-            fieldPath: spec.nodeName
-      image: gcr.io/heptio-images/sonobuoy-plugin-systemd-logs:latest
-      imagePullPolicy: Always
-      name: systemd-logs
-      resources: {}
-      securityContext:
-        privileged: true
-      volumeMounts:
-      - mountPath: /tmp/results
-        name: results
-      - mountPath: /node
-        name: root
-kind: ConfigMap
-metadata:
-  labels:
-    component: sonobuoy
-  name: sonobuoy-plugins-cm
-  namespace: sonobuoy
----
-apiVersion: v1
-kind: Pod
-metadata:
-  labels:
-    component: sonobuoy
-    run: sonobuoy-master
-    tier: analysis
-  name: sonobuoy
-  namespace: sonobuoy
-spec:
-  containers:
-  - command:
-    - /bin/bash
-    - -c
-    - /sonobuoy master --no-exit=true -v 3 --logtostderr
-    env:
-    - name: SONOBUOY_ADVERTISE_IP
-      valueFrom:
-        fieldRef:
-          fieldPath: status.podIP
-    image: sonobuoy/sonobuoy:{{.SonobuoyVersion}}
-    imagePullPolicy: IfNotPresent
-    name: kube-sonobuoy
-    volumeMounts:
-    - mountPath: /etc/sonobuoy
-      name: sonobuoy-config-volume
-    - mountPath: /plugins.d
-      name: sonobuoy-plugins-volume
-    - mountPath: /tmp/sonobuoy
-      name: output-volume
-  restartPolicy: Never
-  serviceAccountName: sonobuoy-serviceaccount
-  tolerations:
-  - key: "kubernetes.io/e2e-evict-taint-key"
-    operator: "Exists"
-  volumes:
-  - configMap:
-      name: sonobuoy-config-cm
-    name: sonobuoy-config-volume
-  - configMap:
-      name: sonobuoy-plugins-cm
-    name: sonobuoy-plugins-volume
-  - emptyDir: {}
-    name: output-volume
----
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    component: sonobuoy
-    run: sonobuoy-master
-  name: sonobuoy-master
-  namespace: sonobuoy
-spec:
-  ports:
-  - port: 8080
-    protocol: TCP
-    targetPort: 8080
-  selector:
-    run: sonobuoy-master
-  type: ClusterIP
